v4.0 A.5: split portability.py — root-last per GRAEAE
Final mega-file split. The 2679-line route file becomes a thin
HTTP shim; domain holds schemas, orchestration, and pure-Python
logic; db holds all SQL.

- mnemos/api/routes/portability.py: thinned to route shim
  (~80 lines: GET /export, POST /import — delegates to domain).
- mnemos/domain/portability/{schemas, export, import_, ids,
  timestamps, serializers, allowlist, version_topology,
  phases}.py: business logic per concern.
- mnemos/db/portability_repo.py: ALL SQL extracted. Phases
  orchestrate; repository functions execute.

No wire-shape changes. preserve_owner=true root-only authority
preserved. Idempotency preserved. 925 tests pass.

Authored-By: Codex
