+++ mnemos/api/routes/portability.py
@@ -1,2679 +1,82 @@
-"""Memory Portability Format (MPF) export / import endpoints.
-
-Reference implementation of `docs/mpf_v0.1.json` (currently 0.1.1).
-
-Scope of this surface:
-
-  * GET  /v1/export — bundles the caller's memories into a single
-    MPF envelope as `kind: memory` records. Non-root callers get
-    only their own owner_id + namespace; root may pass query params
-    to export any owner/namespace/category slice.
-
-    With `?include_sidecars=true` the envelope also carries
-    `kg_triples`, `memory_versions`, and `compression_manifest`
-    sidecar arrays scoped to the same owner/namespace and to the
-    set of memory ids in the export.
-
-  * POST /v1/import — accepts an MPF envelope and upserts
-    `kind: memory` records plus the same three sidecars when
-    present. Non-root rewrites every record's owner_id + namespace
-    to the caller's identity (you can't smuggle other owners' rows
-    in via an import). Root may pass `?preserve_owner=true` to
-    honor the envelope's owner_id + namespace fields verbatim —
-    useful for migrations between MNEMOS instances.
-
-Forward-compat ratchet: the import handler accepts any envelope
-whose mpf_version is `0.1.x`. Newer minor versions can introduce
-optional fields without bumping major; existing field validation
-is unchanged from 0.1.0.
-
-Deferred to later commits:
-
-  * document / fact / event record kinds — currently routed under
-    `unsupported_kinds` on import. Need per-adapter normalization
-    (Graphiti / Cognee mislabel these as payload_version mnemos-3.1
-    instead of mpf-0.1) before they can land in MNEMOS storage.
-  * relations / embeddings sidecars — each is a separate surface
-    with its own round-trip rules.
-  * JSONL streaming for large corpora (single-file JSON only today;
-    tight cap on `limit` keeps request bodies manageable).
-"""
+"""Memory Portability Format (MPF) export / import endpoints."""
 
 from __future__ import annotations
 
-import hashlib
-import json
-import logging
-import uuid
-from datetime import datetime, timezone
-from typing import Any, Dict, List, Optional
+from typing import Annotated, Optional
 
 from fastapi import APIRouter, Body, Depends, HTTPException, Query
-from pydantic import BaseModel, Field
 
 import mnemos.core.lifecycle as _lc
 from mnemos.api.dependencies import UserContext, get_current_user
-from mnemos.core.security import is_root
+from mnemos.domain.portability.export import (
+    _EXPORT_HARD_LIMIT,
+    _EXPORT_SIDECAR_HARD_LIMIT,
+    export_memories as _export_memories,
+)
+from mnemos.domain.portability.import_ import import_memories as _import_memories
+from mnemos.domain.portability.schemas import (
+    ImportStats,
+    MPFEnvelope,
+    MPFRecord,
+    MEMORY_PAYLOAD_VERSION,
+    MPF_VERSION,
+    MPF_VERSION_PREFIX,
+    SOURCE_SYSTEM,
+    SOURCE_VERSION,
+)
+from mnemos.domain.portability.version_topology import _topo_sort_versions
 
-logger = logging.getLogger(__name__)
 router = APIRouter(prefix="/v1", tags=["portability"])
 
-
-# ─── Constants ────────────────────────────────────────────────────────────────
-
-MPF_VERSION = "0.1.1"
-MPF_VERSION_PREFIX = "0.1."  # forward-compat: accept any 0.1.x on import
-MEMORY_PAYLOAD_VERSION = "mnemos-3.1"
-SOURCE_SYSTEM = "mnemos"
-from mnemos._version import __version__ as SOURCE_VERSION
-
-# Server-side export cap. Anything larger should use the streaming
-# JSONL variant — not in this v3.2.0 cut. Prevents a pathological
-# full-table export from pinning memory.
-_EXPORT_HARD_LIMIT = 10_000
-
-# Per-sidecar hard cap. The `limit` parameter only caps the
-# memories rows — without a separate sidecar cap a non-root caller
-# could accumulate a long memory_versions / kg_triples history on
-# a single memory and request `limit=1&include_sidecars=true` to
-# materialize unbounded rows into one JSON envelope (Codex round-40
-# DoS finding). When a sidecar would exceed this cap, the API
-# returns 413 Payload Too Large with guidance to use the CLI
-# streaming exporter, which paginates internally and is the right
-# tool for whole-corpus exports.
-_EXPORT_SIDECAR_HARD_LIMIT = 50_000
-
-
-# ─── Pydantic models (wire shape) ────────────────────────────────────────────
-
-class MPFRecord(BaseModel):
-    """A single record in an MPF envelope. Discriminated union by `kind`."""
-
-    id: str
-    kind: str  # "document" | "memory" | "fact" | "event" (we only emit/accept "memory" today)
-    payload_version: str
-    payload: Dict[str, Any]
-
-
-class MPFEnvelope(BaseModel):
-    """An MPF v0.1.x file envelope.
-
-    Fields kept optional / additive so this endpoint can consume MPF
-    files emitted by other tools (docling, Mem0, Letta) that may
-    populate sidecars this handler doesn't process. Unknown record
-    kinds are skipped per the spec's forward-compatibility rule.
-
-    The three sidecar fields (`kg_triples`, `memory_versions`,
-    `compression_manifest`) are MNEMOS-native and carried as plain
-    dicts to keep the wire form schema-driven (see
-    `docs/mpf_v0.1.json` $defs) rather than tied to the Pydantic
-    model. The handler validates required fields per-row and skips
-    malformed rows during import.
-    """
-
-    mpf_version: str = MPF_VERSION
-    source_system: Optional[str] = SOURCE_SYSTEM
-    source_version: Optional[str] = SOURCE_VERSION
-    source_instance: Optional[str] = None
-    exported_at: Optional[str] = None
-    record_count: Optional[int] = None
-    records: List[MPFRecord] = Field(default_factory=list)
-    kg_triples: Optional[List[Dict[str, Any]]] = None
-    memory_versions: Optional[List[Dict[str, Any]]] = None
-    compression_manifest: Optional[List[Dict[str, Any]]] = None
-
-
-class ImportStats(BaseModel):
-    """Summary of an import run.
-
-    The top-level `imported` / `skipped` / `failed` counters cover
-    `kind: memory` records. The `sidecars_*` dicts break out per-
-    sidecar counts (`kg_triples`, `memory_versions`,
-    `compression_manifest`) so an operator can tell at a glance
-    which surface had write activity. `errors` aggregates per-row
-    failure messages across all surfaces with a `[<surface>]`
-    prefix on sidecar-originating entries.
-    """
-
-    imported: int
-    skipped: int
-    failed: int
-    unsupported_kinds: Dict[str, int] = Field(default_factory=dict)
-    sidecars_imported: Dict[str, int] = Field(default_factory=dict)
-    sidecars_skipped: Dict[str, int] = Field(default_factory=dict)
-    sidecars_failed: Dict[str, int] = Field(default_factory=dict)
-    errors: List[str] = Field(default_factory=list)
-
-
-# ─── Helpers ──────────────────────────────────────────────────────────────────
-
-def _memory_to_record(row) -> MPFRecord:
-    """Shape a memories-row dict into an MPFRecord(kind='memory').
-
-    The payload is the MNEMOS v3.1 native memory schema as-is
-    (content + category + provenance + tenancy fields). An importer
-    running against a different MNEMOS version keys off
-    payload_version to decide what to do with it.
-    """
-    metadata = row.get("metadata") or {}
-    if isinstance(metadata, str):
-        try:
-            metadata = json.loads(metadata)
-        except Exception:
-            metadata = {"_raw": metadata}
-
-    payload: Dict[str, Any] = {
-        "content": row.get("content"),
-        "category": row.get("category"),
-        "subcategory": row.get("subcategory"),
-        "created": _iso(row.get("created")),
-        "updated": _iso(row.get("updated")),
-        "owner_id": row.get("owner_id"),
-        "namespace": row.get("namespace"),
-        "permission_mode": row.get("permission_mode"),
-        "quality_rating": row.get("quality_rating"),
-        "source_model": row.get("source_model"),
-        "source_provider": row.get("source_provider"),
-        "source_session": row.get("source_session"),
-        "source_agent": row.get("source_agent"),
-        "metadata": metadata,
-    }
-    # Strip None entries to keep the envelope tidy — importers default
-    # missing fields via the schema, and nulls on absent columns
-    # inflate envelope size noticeably at 10k rows.
-    payload = {k: v for k, v in payload.items() if v is not None}
-
-    return MPFRecord(
-        id=row["id"],
-        kind="memory",
-        payload_version=MEMORY_PAYLOAD_VERSION,
-        payload=payload,
-    )
-
-
-def _iso(value) -> Optional[str]:
-    """Render a DB timestamp value as an RFC 3339 / ISO 8601 string,
-    or None when the source is None. Top-level helper so the sidecar
-    mappers can share it.
-
-    Naive datetimes are treated as UTC (the MNEMOS convention; the
-    DB never stores a non-UTC value in a TIMESTAMP column). Calling
-    `.astimezone(timezone.utc)` directly on a naive datetime would
-    interpret it as LOCAL TIME and shift it by the host's TZ offset
-    — which means an EDT host's `2026-01-15 10:30:00` would export
-    as `2026-01-15T15:30:00+00:00` and re-import shifted by 5
-    hours. (Codex round-17 finding.) Tag naive values with UTC
-    explicitly before formatting.
-    """
-    if value is None:
-        return None
-    if isinstance(value, str):
-        return value
-    if isinstance(value, datetime):
-        if value.tzinfo is None:
-            return value.replace(tzinfo=timezone.utc).isoformat()
-        return value.astimezone(timezone.utc).isoformat()
-    return str(value)
-
-
-def _kg_triple_to_entry(row) -> Dict[str, Any]:
-    """Convert a `kg_triples` DB row into an MPF kg_triple sidecar
-    entry. The DB stores subject/object as plain TEXT with optional
-    *_type discriminators; we emit them as `subject_literal` /
-    `object_literal` and pass the type tags through verbatim. An
-    importer with a memories table can promote literals back to
-    record references when the literal value matches a known id."""
-    metadata = row.get("metadata") or {}
-    if isinstance(metadata, str):
-        try:
-            metadata = json.loads(metadata)
-        except Exception:
-            metadata = {"_raw": metadata}
-    entry: Dict[str, Any] = {
-        "id": row["id"],
-        "predicate": row["predicate"],
-        "subject_literal": row.get("subject"),
-        "object_literal": row.get("object"),
-        "subject_type": row.get("subject_type"),
-        "object_type": row.get("object_type"),
-        "memory_id": row.get("memory_id"),
-        "confidence": row.get("confidence"),
-        "valid_from": _iso(row.get("valid_from")),
-        "valid_until": _iso(row.get("valid_until")),
-        "created": _iso(row.get("created")),
-        "owner_id": row.get("owner_id"),
-        "namespace": row.get("namespace"),
-    }
-    if metadata:
-        entry["metadata"] = metadata
-    # Prune Nones to keep the envelope tidy. MPF schema marks all
-    # optional fields as nullable so consumers tolerate either form.
-    return {k: v for k, v in entry.items() if v is not None}
-
-
-def _memory_version_to_entry(row) -> Dict[str, Any]:
-    """Convert a `memory_versions` DB row into an MPF
-    memory_version_entry. Carries the full DAG fields (commit_hash,
-    branch, parent_version_id, merge_parents) plus the snapshot's
-    tenancy + provenance so the entry round-trips even if the parent
-    record isn't in the same envelope (partial export case)."""
-    metadata = row.get("metadata") or {}
-    if isinstance(metadata, str):
-        try:
-            metadata = json.loads(metadata)
-        except Exception:
-            metadata = {"_raw": metadata}
-    merge_parents = row.get("merge_parents") or None
-    if merge_parents is not None:
-        # asyncpg renders UUID[] as a list of uuid.UUID; stringify
-        # for the JSON payload.
-        merge_parents = [str(p) for p in merge_parents]
-    parent_version_id = row.get("parent_version_id")
-    if parent_version_id is not None:
-        parent_version_id = str(parent_version_id)
-    entry: Dict[str, Any] = {
-        "id": str(row["id"]),
-        "record_id": row["memory_id"],
-        "version_num": row["version_num"],
-        "commit_hash": row.get("commit_hash"),
-        "branch": row.get("branch"),
-        "parent_version_id": parent_version_id,
-        "merge_parents": merge_parents,
-        "content": row["content"],
-        "category": row.get("category"),
-        "subcategory": row.get("subcategory"),
-        "metadata": metadata or None,
-        "verbatim_content": row.get("verbatim_content"),
-        "owner_id": row.get("owner_id"),
-        "namespace": row.get("namespace"),
-        "permission_mode": row.get("permission_mode"),
-        "source_model": row.get("source_model"),
-        "source_provider": row.get("source_provider"),
-        "source_session": row.get("source_session"),
-        "source_agent": row.get("source_agent"),
-        "snapshot_at": _iso(row.get("snapshot_at")),
-        "snapshot_by": row.get("snapshot_by"),
-        "change_type": row.get("change_type"),
-    }
-    return {k: v for k, v in entry.items() if v is not None}
-
-
-def _compression_variant_to_entry(row) -> Dict[str, Any]:
-    """Convert a `memory_compressed_variants` DB row into an MPF
-    compression_manifest_entry. The DB primary key is `memory_id`
-    (one winner per memory), and `winner_candidate_id` references the
-    contest row that produced this variant."""
-    winner_id = row.get("winner_candidate_id")
-    if winner_id is not None:
-        winner_id = str(winner_id)
-    entry: Dict[str, Any] = {
-        "record_id": row["memory_id"],
-        "engine_id": row["engine_id"],
-        "engine_version": row.get("engine_version"),
-        "compressed_content": row.get("compressed_content"),
-        "compressed_tokens": row.get("compressed_tokens"),
-        "compression_ratio": row.get("compression_ratio"),
-        "quality_score": row.get("quality_score"),
-        "composite_score": row.get("composite_score"),
-        "scoring_profile": row.get("scoring_profile"),
-        "judge_model": row.get("judge_model"),
-        "selected_at": _iso(row.get("selected_at")),
-        "winner_contest_id": winner_id,
-        "owner_id": row.get("owner_id"),
-    }
-    return {k: v for k, v in entry.items() if v is not None}
-
-
-# ─── GET /v1/export ───────────────────────────────────────────────────────────
-
-
-async def _fetch_sidecar(
-    conn,
-    *,
-    table: str,
-    columns: str,
-    memory_id_column: str,
-    memory_ids: List[str],
-    effective_owner: Optional[str],
-    effective_ns: Optional[str],
-    bound_to_memories: bool,
-    null_ok: bool = False,
-    order_by: Optional[str] = None,
-) -> Any:
-    """Build and execute a sidecar SELECT with optional owner /
-    namespace / memory_id filters. Centralizes the placeholder math
-    so each sidecar query stays declarative.
-
-    `bound_to_memories=True` means the sidecar rows must reference an
-    id in `memory_ids`. `null_ok=True` extends that to also include
-    rows whose `memory_id_column` is NULL — used by kg_triples,
-    where some triples are first-class (no memory FK) and some are
-    extracted from a specific memory. With `null_ok=False` (default)
-    only rows with a matching memory_id come back.
-
-    `bound_to_memories=False` drops the memory-id filter entirely —
-    NOT what kg_triples wants in a category-filtered export, since
-    that lets attached triples for non-exported memories slip
-    through (Codex round-5 finding). Use `True + null_ok=True`
-    instead.
-
-    Empty `memory_ids` with `bound_to_memories=True, null_ok=False`
-    short-circuits to no rows without hitting the DB. With
-    `null_ok=True` the query still runs (NULL memory_ids may match).
-    """
-    # Empty memory slice short-circuit. The intent: a category-
-    # filtered export with no matching memories shouldn't leak
-    # all first-class kg_triples in the owner/namespace (round-6
-    # finding). But this short-circuit must NOT fire when the
-    # caller explicitly opted in via null_ok=True — that's the
-    # KG-only migration use case where there are no memories but
-    # the caller does want first-class triples (round-9 finding).
-    if bound_to_memories and not memory_ids and not null_ok:
-        return []
-    conditions: List[str] = []
-    params: List[Any] = []
-    idx = 1
-    if bound_to_memories:
-        if null_ok and memory_ids:
-            conditions.append(
-                f"({memory_id_column} IS NULL OR {memory_id_column} = ANY(${idx}::text[]))"
-            )
-            params.append(memory_ids)
-            idx += 1
-        elif null_ok:
-            # Explicit opt-in with empty slice: first-class only.
-            # Combined with owner/namespace filters this still
-            # scopes the export to the caller's tenancy.
-            conditions.append(f"{memory_id_column} IS NULL")
-        else:
-            conditions.append(f"{memory_id_column} = ANY(${idx}::text[])")
-            params.append(memory_ids)
-            idx += 1
-    if effective_owner:
-        conditions.append(f"owner_id = ${idx}")
-        params.append(effective_owner)
-        idx += 1
-    if effective_ns:
-        conditions.append(f"namespace = ${idx}")
-        params.append(effective_ns)
-        idx += 1
-    where = ("WHERE " + " AND ".join(conditions)) if conditions else ""
-    order = f"ORDER BY {order_by}" if order_by else ""
-    # Append `LIMIT cap+1` so the caller can detect overflow without
-    # materializing more than one row past the limit (Codex round-40).
-    # The over-fetch sentinel is checked at the call site so each
-    # sidecar can produce a surface-specific 413 message.
-    limit_clause = f"LIMIT {_EXPORT_SIDECAR_HARD_LIMIT + 1}"
-    sql = f"SELECT {columns} FROM {table} {where} {order} {limit_clause}"
-    return await conn.fetch(sql, *params)
+__all__ = [
+    "ImportStats", "MEMORY_PAYLOAD_VERSION", "MPFEnvelope", "MPFRecord",
+    "MPF_VERSION", "MPF_VERSION_PREFIX", "SOURCE_SYSTEM", "SOURCE_VERSION",
+    "_EXPORT_HARD_LIMIT", "_EXPORT_SIDECAR_HARD_LIMIT", "_topo_sort_versions",
+    "export_memories", "import_memories", "router",
+]
 
 
 @router.get("/export", response_model=MPFEnvelope)
 async def export_memories(
-    category: Optional[str] = Query(None, description="Filter by category; all categories if unset."),
-    limit: int = Query(1000, ge=1, le=_EXPORT_HARD_LIMIT),
-    offset: int = Query(0, ge=0),
-    owner_id: Optional[str] = Query(None, description="Root only. Export a specific owner's memories; defaults to the caller."),
-    namespace: Optional[str] = Query(None, description="Root only. Export a specific namespace; defaults to the caller's."),
-    include_sidecars: bool = Query(
-        False,
-        description=(
-            "When true, also emit kg_triples / memory_versions / "
-            "compression_manifest sidecars scoped to the same owner "
-            "+ namespace and to the memory ids in the export. "
-            "Defaults false to keep envelope sizes small for the "
-            "common cross-system case."
-        ),
-    ),
-    include_unattached_kg: bool = Query(
-        False,
-        description=(
-            "When true (and include_sidecars=true), also include "
-            "kg_triples whose memory_id IS NULL — first-class facts "
-            "not extracted from a specific memory. Defaults false: "
-            "an export of memories M1, M2 carries only triples "
-            "attached to M1 or M2, NOT every standalone fact in the "
-            "owner/namespace. Set to true for full-corpus exports "
-            "(e.g. cross-system migration) where unattached facts "
-            "are part of the migration scope."
-        ),
-    ),
+    category: Annotated[Optional[str], Query()] = None,
+    limit: Annotated[int, Query(ge=1, le=_EXPORT_HARD_LIMIT)] = 1000,
+    offset: Annotated[int, Query(ge=0)] = 0,
+    owner_id: Annotated[Optional[str], Query()] = None,
+    namespace: Annotated[Optional[str], Query()] = None,
+    include_sidecars: Annotated[bool, Query()] = False,
+    include_unattached_kg: Annotated[bool, Query()] = False,
     user: UserContext = Depends(get_current_user),
 ):
-    """Export memories as an MPF v0.1.x envelope.
-
-    Non-root callers are scoped to their own owner_id + namespace,
-    regardless of the query params. Root callers may target a specific
-    owner/namespace slice for migration or support work.
-
-    When `include_sidecars` is true, the envelope also carries the
-    three MNEMOS-native sidecars (KG triples, memory-version DAG,
-    compression manifest) scoped to the exported memory ids and the
-    same owner/namespace.
-    """
     if not _lc._pool:
         raise HTTPException(status_code=503, detail="Database pool not available")
 
-    if is_root(user):
-        effective_owner = owner_id  # may be None = no filter
-        effective_ns = namespace    # may be None = no filter
-    else:
-        # Non-root cannot exfiltrate outside their own tenancy. If the
-        # caller passed owner/namespace params that don't match their
-        # identity, reject loudly — silent narrowing would hide the
-        # mistake.
-        if owner_id and owner_id != user.user_id:
-            raise HTTPException(status_code=403, detail="cross-owner export requires root")
-        if namespace and namespace != user.namespace:
-            raise HTTPException(status_code=403, detail="cross-namespace export requires root")
-        effective_owner = user.user_id
-        effective_ns = user.namespace
-
-    conditions: List[str] = []
-    params: List[Any] = []
-    idx = 1
-    if effective_owner:
-        conditions.append(f"owner_id = ${idx}")
-        params.append(effective_owner)
-        idx += 1
-    if effective_ns:
-        conditions.append(f"namespace = ${idx}")
-        params.append(effective_ns)
-        idx += 1
-    if category:
-        conditions.append(f"category = ${idx}")
-        params.append(category)
-        idx += 1
-
-    where = ("WHERE " + " AND ".join(conditions)) if conditions else ""
-
-    sql = (
-        "SELECT id, content, category, subcategory, created, updated, "
-        "owner_id, namespace, permission_mode, quality_rating, "
-        "source_model, source_provider, source_session, source_agent, "
-        "metadata "
-        "FROM memories "
-        f"{where} "
-        f"ORDER BY created ASC "
-        f"LIMIT ${idx} OFFSET ${idx + 1}"
-    )
-    params.extend([limit, offset])
-
     async with _lc._pool.acquire() as conn:
-        # Codex round-41 finding: the memories SELECT and the three
-        # sidecar SELECTs must read from a single consistent
-        # snapshot. Under default READ COMMITTED, a concurrent
-        # update between the memories fetch and a memory_versions
-        # fetch could produce an envelope where records[] carries
-        # old content while the sidecar holds the new HEAD —
-        # internally inconsistent for migration backups. Wrap all
-        # reads in a REPEATABLE READ READ ONLY transaction so the
-        # snapshot is taken at the first statement and held for
-        # all subsequent SELECTs.
-        async with conn.transaction(isolation="repeatable_read", readonly=True):
-            rows = await conn.fetch(sql, *params)
-
-            records = [_memory_to_record(dict(r)) for r in rows]
-
-            kg_triples_out: Optional[List[Dict[str, Any]]] = None
-            memory_versions_out: Optional[List[Dict[str, Any]]] = None
-            compression_manifest_out: Optional[List[Dict[str, Any]]] = None
-
-            if include_sidecars:
-                memory_ids = [r["id"] for r in rows]
-
-                # Owner/namespace filters mirror the memories query so a
-                # root caller targeting a single owner gets the matching
-                # slice on each sidecar; non-root is locked to their own
-                # identity by the time we get here. Sidecars also constrain
-                # to the exported memory_ids, so a category-filtered
-                # export only carries the sidecar rows for those memories.
-                def _enforce_sidecar_cap(rows, surface: str):
-                    # Codex round-40 DoS guard: _fetch_sidecar fetches
-                    # cap+1 rows so we can detect overflow without
-                    # materializing more. Bail with 413 here rather
-                    # than building the giant JSON envelope.
-                    if len(rows) > _EXPORT_SIDECAR_HARD_LIMIT:
-                        raise HTTPException(
-                            status_code=413,
-                            detail=(
-                                f"{surface} export exceeds the per-surface "
-                                f"hard limit of {_EXPORT_SIDECAR_HARD_LIMIT} "
-                                "rows for one envelope. Narrow the slice "
-                                "(filter by category, owner_id, namespace, "
-                                "or a smaller `limit`) and re-export, or "
-                                "split the export into multiple chunks."
-                            ),
-                        )
-
-                kg_rows = await _fetch_sidecar(
-                    conn,
-                    table="kg_triples",
-                    columns=(
-                        "id, subject, predicate, object, subject_type, "
-                        "object_type, valid_from, valid_until, memory_id, "
-                        "confidence, created, owner_id, namespace"
-                    ),
-                    memory_id_column="memory_id",
-                    memory_ids=memory_ids,
-                    effective_owner=effective_owner,
-                    effective_ns=effective_ns,
-                    # KG triples have two flavors: first-class (memory_id
-                    # NULL, e.g. external Graphiti-style facts) and
-                    # attached (memory_id pointing at a specific memory
-                    # they were extracted from).
-                    #
-                    # Default behavior (include_unattached_kg=False):
-                    # only attached triples whose memory_id is in the
-                    # exported slice. A category- or limit-filtered
-                    # export of M1+M2 does NOT carry every standalone
-                    # fact in the owner/namespace, which would leak
-                    # unrelated data into a partial export.
-                    #
-                    # Opt-in (include_unattached_kg=True): also include
-                    # the NULL-memory_id rows. Used for full-corpus
-                    # exports / cross-system migrations where first-
-                    # class facts are part of the intended scope.
-                    bound_to_memories=True,
-                    null_ok=include_unattached_kg,
-                )
-                _enforce_sidecar_cap(kg_rows, "kg_triples")
-                kg_triples_out = [_kg_triple_to_entry(dict(r)) for r in kg_rows]
-
-                mv_rows = await _fetch_sidecar(
-                    conn,
-                    table="memory_versions",
-                    columns=(
-                        "id, memory_id, version_num, content, category, "
-                        "subcategory, metadata, verbatim_content, owner_id, "
-                        "namespace, permission_mode, source_model, source_provider, "
-                        "source_session, source_agent, snapshot_at, snapshot_by, "
-                        "change_type, commit_hash, parent_version_id, branch, "
-                        "merge_parents"
-                    ),
-                    memory_id_column="memory_id",
-                    memory_ids=memory_ids,
-                    effective_owner=effective_owner,
-                    effective_ns=effective_ns,
-                    bound_to_memories=True,
-                    # Topological-stable order — parent rows ship before
-                    # child rows so the import-side parent_version_id FK
-                    # check passes even on consumers that import in
-                    # received order. See _import_memory_versions for
-                    # the matching defensive sort. Codex round-8 finding.
-                    order_by="memory_id ASC, branch ASC, version_num ASC",
-                )
-                _enforce_sidecar_cap(mv_rows, "memory_versions")
-                # SQL ORDER BY (memory_id, branch, version_num) is a
-                # heuristic that breaks for forked branches — feature/v1
-                # whose parent_version_id points at main/vN sorts before
-                # main/vN. Apply the real Kahn's-algorithm topo sort
-                # (defined alongside _import_memory_versions) before
-                # emitting, so consumers that import in received order
-                # get a topologically-correct envelope without needing
-                # the defensive sort our own importer applies.
-                memory_versions_out = _topo_sort_versions(
-                    [_memory_version_to_entry(dict(r)) for r in mv_rows]
-                )
-
-                cv_rows = await _fetch_sidecar(
-                    conn,
-                    table="memory_compressed_variants",
-                    columns=(
-                        "memory_id, owner_id, winner_candidate_id, engine_id, "
-                        "engine_version, compressed_content, compressed_tokens, "
-                        "compression_ratio, quality_score, composite_score, "
-                        "scoring_profile, judge_model, selected_at"
-                    ),
-                    memory_id_column="memory_id",
-                    memory_ids=memory_ids,
-                    effective_owner=effective_owner,
-                    # memory_compressed_variants has no `namespace` column;
-                    # tenancy is owner-only here.
-                    effective_ns=None,
-                    bound_to_memories=True,
-                )
-                _enforce_sidecar_cap(cv_rows, "compression_manifest")
-                compression_manifest_out = [_compression_variant_to_entry(dict(r)) for r in cv_rows]
-
-    return MPFEnvelope(
-        mpf_version=MPF_VERSION,
-        source_system=SOURCE_SYSTEM,
-        source_version=SOURCE_VERSION,
-        exported_at=datetime.now(timezone.utc).isoformat(),
-        record_count=len(records),
-        records=records,
-        kg_triples=kg_triples_out,
-        memory_versions=memory_versions_out,
-        compression_manifest=compression_manifest_out,
-    )
-
-
-# ─── POST /v1/import ──────────────────────────────────────────────────────────
-
-
-def _parse_iso(value: Optional[str]) -> Optional[datetime]:
-    """Best-effort parse for the handful of timestamp fields MPF
-    memory payloads carry. Returns None on any failure — the caller
-    lets the DB default fire instead of inserting garbage.
-
-    Returns aware datetime suitable for TIMESTAMPTZ columns. For
-    naive TIMESTAMP columns (memories.created/updated) use
-    `_parse_iso_naive` instead so asyncpg doesn't shift the value
-    by the session's timezone offset (Codex round-17 finding)."""
-    if not value:
-        return None
-    try:
-        # `fromisoformat` handles "2026-01-15T10:30:00+00:00" and its
-        # bare variants. Strip a trailing Z since older pre-3.11
-        # Python doesn't accept it (we're on 3.11+ but belt+braces).
-        if value.endswith("Z"):
-            value = value[:-1] + "+00:00"
-        return datetime.fromisoformat(value)
-    except Exception:
-        return None
-
-
-def _parse_iso_naive(value: Optional[str]) -> Optional[datetime]:
-    """Like _parse_iso, but always returns a NAIVE datetime in UTC.
-
-    For binding into TIMESTAMP (no timezone) columns. asyncpg
-    converts an aware datetime to the SESSION's timezone before
-    stripping tzinfo, which shifts naive TIMESTAMP values by the
-    server's TZ offset. Pre-strip to UTC-as-naive to bypass that.
-    """
-    parsed = _parse_iso(value)
-    if parsed is None:
-        return None
-    if parsed.tzinfo is not None:
-        parsed = parsed.astimezone(timezone.utc).replace(tzinfo=None)
-    return parsed
-
-
-async def _build_referenced_memory_allowlist(
-    conn,
-    envelope: "MPFEnvelope",
-    *,
-    scope_owner: Optional[str] = None,
-    scope_namespace: Optional[str] = None,
-) -> Dict[str, tuple]:
-    """Return {memory_id: (owner_id, namespace)} for every sidecar-
-    referenced memory_id, looked up in DB.
-
-    When `scope_owner` and/or `scope_namespace` are provided, the
-    SELECT is filtered to that tenancy. Non-root callers always
-    pass their own identity here so the lookup is structurally
-    scoped — a foreign memory_id simply doesn't appear in the
-    result, indistinguishable from a non-existent id (Codex
-    round-13 finding: an unscoped lookup combined with a
-    descriptive rejection error lets a non-root caller probe
-    foreign tenants' memory_ids and learn their owners).
-
-    Root callers with preserve_owner=true pass scope_owner=None
-    so the migration/admin path can resolve any memory_id.
-
-    First-class kg_triples (memory_id absent) are not subject to
-    this check.
-    """
-    referenced: set = set()
-    for entry in envelope.kg_triples or []:
-        mid = entry.get("memory_id")
-        if mid:
-            referenced.add(mid)
-    for entry in envelope.memory_versions or []:
-        rid = entry.get("record_id")
-        if rid:
-            referenced.add(rid)
-    for entry in envelope.compression_manifest or []:
-        rid = entry.get("record_id")
-        if rid:
-            referenced.add(rid)
-    if not referenced:
-        return {}
-    sql = "SELECT id, owner_id, namespace FROM memories WHERE id = ANY($1::text[])"
-    params: List[Any] = [list(referenced)]
-    if scope_owner is not None:
-        sql += " AND owner_id = $2"
-        params.append(scope_owner)
-        if scope_namespace is not None:
-            sql += " AND namespace = $3"
-            params.append(scope_namespace)
-    elif scope_namespace is not None:
-        sql += " AND namespace = $2"
-        params.append(scope_namespace)
-    rows = await conn.fetch(sql, *params)
-    return {r["id"]: (r["owner_id"], r["namespace"]) for r in rows}
-
-
-def _is_allowed_reference(
-    memory_id: Optional[str],
-    *,
-    effective_owner: str,
-    effective_namespace: Optional[str],
-    allowlist: Dict[str, tuple],
-    require_namespace_match: bool = True,
-) -> tuple[bool, str]:
-    """Return (allowed, reason). The sidecar is allowed iff the
-    referenced memory_id either (a) is None — first-class triple
-    case — or (b) exists in the allowlist AND the memory's actual
-    owner+namespace matches the post-rewrite effective owner+ns
-    that this sidecar would land under.
-
-    `require_namespace_match=False` for compression_manifest, which
-    has no namespace column — owner match is sufficient there.
-    """
-    if memory_id is None:
-        return True, ""
-    if memory_id not in allowlist:
-        # Generic message for both nonexistent and foreign-tenant
-        # ids. The caller's allowlist SELECT is structurally scoped
-        # to their tenancy (see _build_referenced_memory_allowlist),
-        # so the two cases are indistinguishable from here — which
-        # is the point: a non-root attacker can't probe for foreign
-        # memory_ids via rejection-message inference.
-        return False, (
-            f"record_id {memory_id!r} not in caller-owned memory id set; "
-            "skipped (cross-tenant attachment refused)"
+        return await _export_memories(
+            conn,
+            user=user,
+            category=category,
+            limit=limit,
+            offset=offset,
+            owner_id=owner_id,
+            namespace=namespace,
+            include_sidecars=include_sidecars,
+            include_unattached_kg=include_unattached_kg,
         )
-    actual_owner, actual_ns = allowlist[memory_id]
-    # The remaining branches only fire for root callers with
-    # preserve_owner=true (the unscoped allowlist path). Their
-    # descriptive messages are intentional — root is doing
-    # migration work and needs the diagnostic to debug envelope
-    # tenancy mismatches.
-    if actual_owner != effective_owner:
-        return False, (
-            f"record_id {memory_id!r} belongs to owner {actual_owner!r}, "
-            f"not the sidecar's effective owner {effective_owner!r}; "
-            "skipped (cross-tenant attachment refused)"
-        )
-    if require_namespace_match and actual_ns != effective_namespace:
-        return False, (
-            f"record_id {memory_id!r} is in namespace {actual_ns!r}, "
-            f"not the sidecar's effective namespace {effective_namespace!r}; "
-            "skipped (cross-tenant attachment refused)"
-        )
-    return True, ""
-
-
-def _derive_caller_scoped_uuid(
-    envelope_id: str,
-    *,
-    caller_owner: str,
-    caller_namespace: str,
-    extra: str = "",
-) -> str:
-    """Derive a deterministic UUID from caller identity + envelope
-    id (+ optional extra payload for content-binding).
-
-    Used for sidecar primary keys (memory_versions.id,
-    kg_triples.id) under non-root imports. Without this,
-    sidecar PKs are envelope-supplied UUIDs that collide globally
-    when the same envelope is imported by multiple non-root callers
-    (Codex round-15 finding). Caller-scoping makes each importer's
-    sidecar id space disjoint, identical to the records-loop fix.
-
-    Returns a string-formatted UUID (uuid.UUID(...)). The hash is
-    sha1-based via uuid.uuid5; collision odds are negligible for
-    the input scale.
-    """
-    namespace = uuid.UUID("6ba7b810-9dad-11d1-80b4-00c04fd430c8")  # uuid.NAMESPACE_DNS
-    name = "\x00".join([caller_owner, caller_namespace, envelope_id, extra])
-    return str(uuid.uuid5(namespace, name))
-
-
-def _derive_caller_scoped_id(
-    envelope_id: str,
-    *,
-    caller_owner: str,
-    caller_namespace: str,
-    content: str,
-) -> str:
-    """Derive a caller-scoped, deterministic id for non-root
-    record imports. Replaces the envelope's record.id so a
-    non-root caller cannot probe for foreign memory_ids via
-    `imported` vs `skipped` outcome inference (Codex round-14
-    finding).
-
-    Hash inputs: caller identity + envelope id + content. Mixing
-    in caller identity ensures alice and carol can both have
-    identical content without colliding (different hashes).
-    Mixing in the envelope id keeps re-imports of the SAME
-    envelope idempotent (same id derives same hash → ON CONFLICT
-    DO NOTHING). Mixing in content prevents an adversary from
-    pre-computing a target id space (they'd need to know the
-    target's content).
-
-    Returns a string that fits in the memories.id TEXT column.
-    """
-    h = hashlib.sha256(
-        b"\x00".join([
-            caller_owner.encode("utf-8"),
-            caller_namespace.encode("utf-8"),
-            envelope_id.encode("utf-8"),
-            content.encode("utf-8"),
-        ])
-    ).hexdigest()[:32]
-    return f"mnemos_{h}"
-
-
-def _row_owner_ns(
-    entry: Dict[str, Any],
-    *,
-    caller_user_id: str,
-    caller_namespace: str,
-    preserve_owner: bool,
-    has_namespace_column: bool = True,
-) -> tuple[str, Optional[str]]:
-    """Apply the same owner/namespace rewrite rule to a sidecar row
-    that the records loop applies to memories. Non-root or
-    preserve_owner=false → caller identity. preserve_owner=true →
-    honor the entry's fields, falling back to caller identity when
-    the entry omits them.
-
-    `has_namespace_column=False` for tables without a namespace
-    column (compression manifest); the second element is None.
-    """
-    if preserve_owner:
-        owner = entry.get("owner_id") or caller_user_id
-        ns = (entry.get("namespace") or caller_namespace) if has_namespace_column else None
-    else:
-        owner = caller_user_id
-        ns = caller_namespace if has_namespace_column else None
-    return owner, ns
-
-
-def _bump(counter: Dict[str, int], key: str) -> None:
-    counter[key] = counter.get(key, 0) + 1
-
-
-async def _import_kg_triples(
-    conn,
-    sidecar: List[Dict[str, Any]],
-    *,
-    caller_user_id: str,
-    caller_namespace: str,
-    preserve_owner: bool,
-    stats: ImportStats,
-    allowlist: Dict[str, tuple],
-    inserted_record_ids: Optional[set] = None,
-) -> None:
-    """Upsert MPF kg_triples sidecar entries into the kg_triples
-    table. Idempotent on `id`. Subject/object literals and id
-    references both flatten into the DB's TEXT columns; the
-    *_type tags persist as discriminators.
-    """
-    surface = "kg_triples"
-    # Caller-scoped PK rewrite for non-root imports — same as the
-    # memory_versions helper (Codex round-15 finding). kg_triples
-    # has no FK self-references, so we only remap entry["id"]
-    # itself; subject_id / object_id are free-form references
-    # that the consumer can interpret.
-    if not preserve_owner:
-        for entry in sidecar:
-            eid = entry.get("id")
-            if eid:
-                entry["id"] = _derive_caller_scoped_uuid(
-                    str(eid),
-                    caller_owner=caller_user_id,
-                    caller_namespace=caller_namespace,
-                    extra="kg_triples",
-                )
-
-    for entry in sidecar:
-        if not entry.get("id") or not entry.get("predicate"):
-            _bump(stats.sidecars_failed, surface)
-            stats.errors.append(f"[{surface}] missing required id/predicate; skipped")
-            continue
-        # Per the schema's anyOf, exactly one of subject_id /
-        # subject_literal must be present. Prefer the literal because
-        # the DB has a single TEXT column; downgrade subject_id to
-        # literal form (the importer can't re-resolve a foreign id).
-        subject = entry.get("subject_literal") or entry.get("subject_id")
-        obj = entry.get("object_literal") or entry.get("object_id")
-        if not subject:
-            _bump(stats.sidecars_failed, surface)
-            stats.errors.append(
-                f"[{surface}] {entry['id']}: missing subject_literal/subject_id; skipped"
-            )
-            continue
-        if not obj:
-            # Object can be missing in some adapters; default to empty
-            # string rather than failing the row outright.
-            obj = ""
-        row_owner, row_ns = _row_owner_ns(
-            entry,
-            caller_user_id=caller_user_id,
-            caller_namespace=caller_namespace,
-            preserve_owner=preserve_owner,
-        )
-        # Tenant-scope check: the triple's memory_id (when present)
-        # must reference a memory whose actual owner+namespace match
-        # the post-rewrite owner+ns we'd persist this triple under.
-        # Blocks alice from attaching kg_triples to bob's memory_id.
-        allowed, reason = _is_allowed_reference(
-            entry.get("memory_id"),
-            effective_owner=row_owner,
-            effective_namespace=row_ns,
-            allowlist=allowlist,
-        )
-        if not allowed:
-            _bump(stats.sidecars_failed, surface)
-            stats.errors.append(f"[{surface}] {entry['id']}: {reason}")
-            continue
-        # Per-row SAVEPOINT — a Postgres error on one kg_triple
-        # mustn't abort the surrounding import transaction.
-        try:
-            async with conn.transaction():
-                row = await conn.execute(
-                    """
-                    INSERT INTO kg_triples (
-                        id, subject, predicate, object,
-                        subject_type, object_type,
-                        valid_from, valid_until,
-                        memory_id, confidence, created,
-                        owner_id, namespace
-                    )
-                    VALUES (
-                        $1, $2, $3, $4,
-                        $5, $6,
-                        COALESCE($7, NOW()), $8,
-                        $9, COALESCE($10, 1.0),
-                        COALESCE($11, NOW()),
-                        $12, $13
-                    )
-                    ON CONFLICT (id) DO NOTHING
-                    """,
-                    entry["id"], subject, entry["predicate"], obj,
-                    entry.get("subject_type"), entry.get("object_type"),
-                    _parse_iso(entry.get("valid_from")),
-                    _parse_iso(entry.get("valid_until")),
-                    entry.get("memory_id"),
-                    entry.get("confidence"),
-                    _parse_iso(entry.get("created")),
-                    row_owner, row_ns,
-                )
-            if row == "INSERT 0 0":
-                # Conflict-skip: verify the existing row matches
-                # the envelope's claim. kg_triples.memory_id has no
-                # FK so prior-lifetime / orphan rows can survive a
-                # memory delete, and the round-23 pattern (exact
-                # match on every column) applies here too —
-                # otherwise re-imports of corrected triples
-                # silently leave stale facts in place. Codex
-                # round-30 finding.
-                existing = await conn.fetchrow(
-                    "SELECT subject, predicate, object, subject_type, "
-                    "object_type, memory_id, confidence, owner_id, "
-                    "namespace, valid_from, valid_until, created "
-                    "FROM kg_triples WHERE id = $1",
-                    entry["id"],
-                )
-                expected_valid_from = _parse_iso(entry.get("valid_from"))
-                expected_valid_until = _parse_iso(entry.get("valid_until"))
-                expected_created = _parse_iso(entry.get("created"))
-                # COALESCE-tolerance for INSERT-time NOW() defaults
-                # (Codex round-38). valid_from and created use
-                # COALESCE($N, NOW()) on INSERT — when envelope omits
-                # them, DB stores NOW(). On idempotent retry the
-                # envelope still has None, so strict equality
-                # (existing != None) would falsely reject the retry.
-                # valid_until is plain $N (no COALESCE) so strict
-                # NULL=NULL comparison stays correct there.
-                #
-                # Round-39 escalation: tolerance is ONLY safe when
-                # the referenced memory was NOT just inserted in
-                # this transaction. For a fresh memory, an
-                # INSERT 0 0 on kg_triples cannot be an idempotent
-                # retry — kg_triples lacks an FK to memories so a
-                # stale prior-lifetime row can survive deletion of
-                # its memory. Without the gate, the omitted-NOW()
-                # tolerance would let that stale row attach to
-                # the freshly inserted memory.
-                referenced_memory_id = entry.get("memory_id")
-                fresh_memory = (
-                    inserted_record_ids is not None
-                    and referenced_memory_id is not None
-                    and referenced_memory_id in inserted_record_ids
-                )
-                tolerate_valid_from = (
-                    not fresh_memory and entry.get("valid_from") is None
-                )
-                tolerate_created = (
-                    not fresh_memory and entry.get("created") is None
-                )
-                if existing is None or (
-                    existing["subject"] != subject
-                    or existing["predicate"] != entry["predicate"]
-                    or existing["object"] != obj
-                    or existing["subject_type"] != entry.get("subject_type")
-                    or existing["object_type"] != entry.get("object_type")
-                    or existing["memory_id"] != entry.get("memory_id")
-                    # confidence: INSERT uses COALESCE($N, 1.0)
-                    # so an omitted value persists as 1.0 — compare
-                    # against that default explicitly. Round-34 fix.
-                    or existing["confidence"] != (
-                        entry["confidence"] if entry.get("confidence") is not None else 1.0
-                    )
-                    or existing["owner_id"] != row_owner
-                    or existing["namespace"] != row_ns
-                    # Temporal columns (round-31): a corrected
-                    # temporal edge with the same id but different
-                    # valid_from / valid_until / created window
-                    # must NOT be counted as skipped. When the
-                    # envelope explicitly supplied a temporal value
-                    # we compare strictly. When envelope omitted it
-                    # AND the column uses COALESCE(NOW()) on insert,
-                    # tolerate the DB-generated value so retries are
-                    # idempotent (round-38).
-                    or (not tolerate_valid_from and existing["valid_from"] != expected_valid_from)
-                    or existing["valid_until"] != expected_valid_until
-                    or (not tolerate_created and existing["created"] != expected_created)
-                ):
-                    _bump(stats.sidecars_failed, surface)
-                    stats.errors.append(
-                        f"[{surface}] {entry['id']}: existing row "
-                        "doesn't match envelope claim (likely "
-                        "stale or orphaned triple); rejected"
-                    )
-                    continue
-                _bump(stats.sidecars_skipped, surface)
-            else:
-                _bump(stats.sidecars_imported, surface)
-        except Exception as exc:
-            _bump(stats.sidecars_failed, surface)
-            stats.errors.append(f"[{surface}] {entry['id']}: {type(exc).__name__}: {exc}")
-            logger.exception("MPF kg_triples import failed for entry %s", entry.get("id"))
-
-
-async def _restore_memory_branches(
-    conn,
-    memory_ids: List[str],
-    *,
-    authorized_version_uuids: Optional[List[str]] = None,
-) -> None:
-    """After memory_versions sidecar import, repopulate memory_branches
-    HEAD pointers for the imported records.
-
-    The mnemos_version_snapshot trigger normally upserts memory_branches
-    on every memory INSERT, but during a CHARON import the trigger is
-    suppressed via the mnemos.suppress_version_snapshot GUC. Without
-    this restore step, /log + branch-walk endpoints see the version
-    rows but no HEAD pointer, so DAG queries return empty.
-
-    `authorized_version_uuids` scopes the SELECT to ONLY the version
-    rows that were imported / authorized in THIS request — pre-existing
-    stale rows from a prior lifetime (e.g. from a memory that was
-    deleted and is now being re-imported) DO NOT participate in HEAD
-    selection. Without this scope, an attacker could re-import a
-    deterministic id and have the prior lifetime's main HEAD survive
-    into the new memory's branch state. (Codex round-21 finding.)
-    Pass None to fall back to all-DB-rows behavior for callers that
-    don't have a sidecar-derived authorized set.
-    """
-    if not memory_ids:
-        return
-    if authorized_version_uuids is not None and not authorized_version_uuids:
-        return
-    if authorized_version_uuids is not None:
-        rows = await conn.fetch(
-            """
-            SELECT DISTINCT ON (memory_id, branch)
-                memory_id, branch, id AS head_version_id
-            FROM memory_versions
-            WHERE memory_id = ANY($1::text[])
-              AND id = ANY($2::uuid[])
-            ORDER BY memory_id, branch, version_num DESC
-            """,
-            memory_ids, authorized_version_uuids,
-        )
-    else:
-        rows = await conn.fetch(
-            """
-            SELECT DISTINCT ON (memory_id, branch)
-                memory_id, branch, id AS head_version_id
-            FROM memory_versions
-            WHERE memory_id = ANY($1::text[])
-            ORDER BY memory_id, branch, version_num DESC
-            """,
-            memory_ids,
-        )
-    for r in rows:
-        await conn.execute(
-            """
-            INSERT INTO memory_branches (memory_id, name, head_version_id, created_by)
-            VALUES ($1, $2, $3, NULL)
-            ON CONFLICT (memory_id, name) DO UPDATE
-            SET head_version_id = EXCLUDED.head_version_id
-            """,
-            r["memory_id"], r["branch"], r["head_version_id"],
-        )
-
-
-async def _validate_version_parents(
-    conn,
-    parent_uuids: List[str],
-    *,
-    expected_record_id: str,
-    effective_owner: str,
-    effective_ns: Optional[str],
-    in_envelope_index: Dict[str, Dict[str, Any]],
-    preserve_owner: bool = True,
-    require_in_envelope: bool = False,
-    freshly_inserted_uuids: Optional[set] = None,
-) -> tuple:
-    """Verify every parent UUID points at a row under the same
-    record_id + owner + namespace as the entry being imported.
-
-    DB IS THE SOURCE OF TRUTH. The in-envelope index alone is not
-    enough — INSERT ... ON CONFLICT (id) DO NOTHING means an
-    envelope-supplied "parent" entry with an id that already exists
-    in DB will be SKIPPED. The pre-existing DB row's tenancy is
-    what the child's FK actually resolves to. So if an adversary
-    crafts a fake parent entry labeled with their tenancy but
-    using a known-foreign UUID, the conflict-skip path bypasses
-    the in-envelope check and the child links to the foreign row.
-    (Codex round-12 finding.)
-
-    Algorithm:
-      1. SELECT every parent UUID from memory_versions in DB.
-      2. For UUIDs found in DB: DB tenancy is authoritative;
-         reject if it doesn't match the entry's effective owner+ns
-         + record_id.
-      3. For UUIDs only in the envelope (truly new parents):
-         the envelope's claim is checked; topological sort guarantees
-         these will be inserted before children, so once persisted
-         the DB row will reflect the envelope's claim.
-      4. UUIDs in neither DB nor envelope: reject (dangling).
-
-    Returns (all_valid, bad_parents_list).
-    """
-    if not parent_uuids:
-        return True, []
-    # DB-truth lookup over every parent UUID — including those
-    # also referenced by the envelope. The envelope's claim only
-    # matters when the UUID is GENUINELY new (not in DB yet).
-    rows = await conn.fetch(
-        "SELECT id::text AS id, memory_id, owner_id, namespace "
-        "FROM memory_versions WHERE id = ANY($1::uuid[])",
-        parent_uuids,
-    )
-    db_truth: Dict[str, tuple] = {
-        r["id"]: (r["memory_id"], r["owner_id"], r["namespace"])
-        for r in rows
-    }
-    bad: List[str] = []
-    for p in parent_uuids:
-        if require_in_envelope:
-            # Parent must be FRESHLY INSERTED in this transaction —
-            # not just present in in_envelope_index. An in-envelope
-            # parent entry can ON CONFLICT skip onto a stale prior-
-            # lifetime row (Codex round-29 finding), which the
-            # equality check passes but the row is NOT a row this
-            # transaction wrote. The freshly_inserted_uuids set is
-            # populated only when an INSERT actually returned
-            # INSERT 0 1, never when ON CONFLICT skipped.
-            #
-            # Topo sort guarantees parents come before children,
-            # so when this child is being validated, the parent
-            # has already been processed and (if legitimately
-            # freshly inserted) is in the set.
-            if freshly_inserted_uuids is None or p not in freshly_inserted_uuids:
-                bad.append(p)
-                continue
-        if p in db_truth:
-            # DB has this parent. Tenancy is whatever's in DB,
-            # NOT what the envelope claims (the envelope's INSERT
-            # would no-op via ON CONFLICT). Validate against DB.
-            mem_id, owner, ns = db_truth[p]
-            if (mem_id != expected_record_id
-                    or owner != effective_owner
-                    or ns != effective_ns):
-                bad.append(p)
-            continue
-        if p in in_envelope_index:
-            ref = in_envelope_index[p]
-            ref_record = ref.get("record_id")
-            same_record = ref_record == expected_record_id
-            if preserve_owner:
-                # Root + preserve_owner=true: envelope owner/ns
-                # are honored on insert. Validate against them.
-                ref_owner = ref.get("owner_id")
-                ref_ns = ref.get("namespace")
-                same_owner = (ref_owner is None) or (ref_owner == effective_owner)
-                same_ns = (ref_ns is None) or (ref_ns == effective_ns)
-            else:
-                # Non-root: every in-envelope parent will be
-                # persisted under the caller's identity by the
-                # owner-rewrite, so its post-INSERT owner/ns ARE
-                # the caller's. The envelope's raw owner/ns fields
-                # are about to be overwritten — don't validate
-                # against them. Only the same-record_id constraint
-                # matters here. Codex round-15 finding: previous
-                # code rejected legitimate non-root cross-system
-                # imports because bob's exported envelope carries
-                # owner=bob in sidecar entries, but alice will
-                # land them under owner=alice.
-                same_owner = True
-                same_ns = True
-            if not (same_record and same_owner and same_ns):
-                bad.append(p)
-            continue
-        # Not in DB, not in envelope — dangling reference.
-        bad.append(p)
-    return (not bad), bad
-
-
-def _topo_sort_versions(sidecar: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
-    """Kahn's-algorithm topological sort over a memory_versions
-    sidecar list, treating parent_version_id + merge_parents as
-    incoming edges.
-
-    Parents that aren't in the sidecar (already in DB, or absent
-    entirely) count as zero in-degree edges — they're treated as
-    roots from this envelope's POV; their FK will resolve at INSERT
-    time against existing DB rows.
-
-    Tie-breaker: lower version_num first, then id-string ASC, so
-    the order is deterministic for round-trip reproducibility.
-
-    Cycles (shouldn't happen in a real DAG but defensive): any
-    nodes left in the working set after Kahn's terminates get
-    appended at the end in tie-broken order. The DB will reject
-    them at INSERT time, but the sort itself never deadlocks.
-    """
-    if not sidecar:
-        return []
-
-    # Build the index: id → entry. Some entries may have no id
-    # (already counted as failed by the caller's required-field
-    # check); skip those for the graph but keep them at the end
-    # so the loop still sees them and counts the failure.
-    by_id: Dict[str, Dict[str, Any]] = {}
-    no_id: List[Dict[str, Any]] = []
-    for entry in sidecar:
-        eid = entry.get("id")
-        if eid:
-            by_id[str(eid)] = entry
-        else:
-            no_id.append(entry)
-
-    # Build incoming-edge counts. Only edges whose target is also
-    # in this envelope contribute.
-    in_degree: Dict[str, int] = {eid: 0 for eid in by_id}
-    children: Dict[str, List[str]] = {eid: [] for eid in by_id}
-    for eid, entry in by_id.items():
-        parents: List[str] = []
-        pv = entry.get("parent_version_id")
-        if pv:
-            parents.append(str(pv))
-        for mp in entry.get("merge_parents") or []:
-            if mp:
-                parents.append(str(mp))
-        for p in parents:
-            if p in by_id:
-                in_degree[eid] += 1
-                children[p].append(eid)
-
-    def _key(eid: str) -> tuple:
-        e = by_id[eid]
-        return (int(e.get("version_num") or 0), eid)
-
-    ready = sorted([eid for eid, d in in_degree.items() if d == 0], key=_key)
-    out: List[Dict[str, Any]] = []
-    while ready:
-        eid = ready.pop(0)
-        out.append(by_id[eid])
-        for child in children[eid]:
-            in_degree[child] -= 1
-            if in_degree[child] == 0:
-                # Insert maintaining tie-break order.
-                k = _key(child)
-                lo, hi = 0, len(ready)
-                while lo < hi:
-                    mid = (lo + hi) // 2
-                    if _key(ready[mid]) < k:
-                        lo = mid + 1
-                    else:
-                        hi = mid
-                ready.insert(lo, child)
-
-    if len(out) < len(by_id):
-        # Cycle detected (or unreachable dep graph). Append the
-        # remaining nodes in tie-broken order; the DB will reject
-        # them at INSERT time, which the per-row SAVEPOINT and
-        # all-or-nothing post-check both handle.
-        leftover = sorted(
-            [eid for eid in by_id if by_id[eid] not in out],
-            key=_key,
-        )
-        out.extend(by_id[eid] for eid in leftover)
-
-    return out + no_id
-
-
-async def _import_memory_versions(
-    conn,
-    sidecar: List[Dict[str, Any]],
-    *,
-    caller_user_id: str,
-    caller_namespace: str,
-    preserve_owner: bool,
-    stats: ImportStats,
-    allowlist: Dict[str, tuple],
-    inserted_record_ids: Optional[set] = None,
-) -> tuple:
-    """Upsert MPF memory_version_entry sidecar entries into
-    memory_versions. Idempotent on `id`. Carries the full DAG
-    triplet (commit_hash, parent_version_id, branch); merge_parents
-    flows through verbatim as a UUID[].
-
-    Returns ``(authorized_record_ids, failed_record_ids)``:
-      - authorized_record_ids: record_ids whose version row was
-        successfully imported OR was already present (skipped via
-        ON CONFLICT). Used downstream for memory_branches HEAD
-        restoration — only memory ids the caller was actually
-        authorized to write get touched.
-      - failed_record_ids: record_ids that had at least one entry
-        rejected (allowlist failure, validation failure, DB error,
-        missing required field). The caller uses this to enforce
-        the all-or-nothing-history contract: if a record was
-        INSERTed in this request AND any of its memory_versions
-        entries failed, the import must roll back. Otherwise the
-        record commits with partial / inconsistent history.
-    """
-    surface = "memory_versions"
-    authorized_record_ids: set = set()
-    failed_record_ids: set = set()
-    # Set of memory_versions.id UUIDs that successfully landed in
-    # THIS request's transaction (or hit ON CONFLICT against a
-    # row that already represented this exact version under the
-    # caller's tenancy). Used to scope _restore_memory_branches
-    # so prior-lifetime stale rows can't be selected as HEAD.
-    authorized_version_uuids: set = set()
-    # Subset of authorized_version_uuids: UUIDs that THIS request
-    # actually INSERTed (not ON CONFLICT skipped against a pre-
-    # existing row). Used for the round-29 stale-via-conflict
-    # bypass guard: parents of newly-inserted records must come
-    # from this set, not just from in_envelope_index, because an
-    # in-envelope parent entry can ON CONFLICT skip onto a stale
-    # prior-lifetime row.
-    freshly_inserted_version_uuids: set = set()
-
-    # Caller-scoped PK rewrite for non-root imports. Each entry's
-    # `id` plus every parent_version_id/merge_parents reference
-    # gets translated through this map so two different non-root
-    # callers importing the same envelope land in disjoint id
-    # spaces (Codex round-15 finding). Root + preserve_owner=true
-    # keeps envelope ids verbatim — the admin migration path.
-    non_root_pk_rewrite = not (preserve_owner)
-    version_id_remap: Dict[str, str] = {}
-    if non_root_pk_rewrite:
-        for entry in sidecar:
-            eid = entry.get("id")
-            if eid:
-                version_id_remap[str(eid)] = _derive_caller_scoped_uuid(
-                    str(eid),
-                    caller_owner=caller_user_id,
-                    caller_namespace=caller_namespace,
-                    extra="memory_versions",
-                )
-        # Translate the entry's own id and any parent references.
-        # Also remap commit_hash — it has a global UNIQUE index in
-        # memory_versions, so two non-root callers importing the
-        # same envelope would collide there too (Codex round-16
-        # finding). Caller-scope by hashing (caller, raw hash):
-        # uniqueness preserved across callers, traceability preserved
-        # within a caller's space.
-        for entry in sidecar:
-            if entry.get("id") in version_id_remap:
-                entry["id"] = version_id_remap[entry["id"]]
-            pv = entry.get("parent_version_id")
-            if pv and str(pv) in version_id_remap:
-                entry["parent_version_id"] = version_id_remap[str(pv)]
-            mp_list = entry.get("merge_parents")
-            if mp_list:
-                entry["merge_parents"] = [
-                    version_id_remap.get(str(mp), str(mp)) for mp in mp_list
-                ]
-            ch = entry.get("commit_hash")
-            if ch:
-                entry["commit_hash"] = hashlib.sha256(
-                    b"\x00".join([
-                        caller_user_id.encode("utf-8"),
-                        caller_namespace.encode("utf-8"),
-                        str(ch).encode("utf-8"),
-                    ])
-                ).hexdigest()
-
-    # Real topological sort over parent_version_id + merge_parents
-    # so parent rows are inserted before children regardless of
-    # envelope order or branch name. The previous (record_id,
-    # branch, version_num) sort failed for forked histories where
-    # a feature branch's v1 parents to main's vN — alphabetic on
-    # branch name put feature/v1 before main/vN, FK violation
-    # (Codex round-9 finding).
-    #
-    # Algorithm: Kahn's. Build edges child→parent for both
-    # parent_version_id and merge_parents (UUID[]). A node's
-    # "in-degree" is the number of its parents that ALSO appear
-    # in this envelope; parents already in DB count as zero
-    # (treated as roots, their FK will resolve at INSERT time).
-    # Pop zero-in-degree nodes in ties broken by version_num.
-    sidecar = _topo_sort_versions(sidecar)
-    # Index in-envelope entries by id for parent-validation
-    # lookups. parent_version_id and merge_parents may point at
-    # entries inside this envelope OR at rows already in DB; the
-    # validator falls back to a SELECT for the latter.
-    in_envelope_index = {
-        str(e["id"]): e for e in sidecar if e.get("id")
-    }
-    for entry in sidecar:
-        # Capture record_id even when other required fields are
-        # missing — we need it for the failure-tracking set so the
-        # caller can decide whether to roll back.
-        record_id_for_tracking = entry.get("record_id")
-        for required in ("id", "record_id", "version_num", "content"):
-            if entry.get(required) in (None, ""):
-                _bump(stats.sidecars_failed, surface)
-                stats.errors.append(
-                    f"[{surface}] missing required field {required!r}; skipped"
-                )
-                if record_id_for_tracking:
-                    failed_record_ids.add(record_id_for_tracking)
-                break
-        else:
-            row_owner, row_ns = _row_owner_ns(
-                entry,
-                caller_user_id=caller_user_id,
-                caller_namespace=caller_namespace,
-                preserve_owner=preserve_owner,
-            )
-            # Same tenant-scope check as kg_triples — every version
-            # must attach to a memory the caller (post-rewrite) owns.
-            allowed, reason = _is_allowed_reference(
-                entry.get("record_id"),
-                effective_owner=row_owner,
-                effective_namespace=row_ns,
-                allowlist=allowlist,
-            )
-            if not allowed:
-                _bump(stats.sidecars_failed, surface)
-                stats.errors.append(f"[{surface}] {entry['id']}: {reason}")
-                failed_record_ids.add(entry["record_id"])
-                continue
-
-            # Tenant-scope every parent UUID. parent_version_id and
-            # merge_parents both reference memory_versions(id) — the
-            # DB FK only proves the UUID exists, NOT that it points
-            # at a same-memory, same-tenant row. Validate explicitly
-            # so an adversarial sidecar can't link its child commit
-            # to another tenant's parent (Codex round-11 finding).
-            parent_uuids: List[str] = []
-            if entry.get("parent_version_id"):
-                parent_uuids.append(str(entry["parent_version_id"]))
-            for mp in entry.get("merge_parents") or []:
-                if mp:
-                    parent_uuids.append(str(mp))
-            if parent_uuids:
-                # If this version's record_id was newly INSERTed
-                # in this request, require its parents to be in
-                # the current envelope (not stale DB rows from a
-                # prior lifetime). If the record_id pre-existed
-                # (ON CONFLICT), DB-only parents are legitimate
-                # — the record's history was already valid.
-                require_in_envelope = (
-                    inserted_record_ids is not None
-                    and entry["record_id"] in inserted_record_ids
-                )
-                ok, bad = await _validate_version_parents(
-                    conn, parent_uuids,
-                    expected_record_id=entry["record_id"],
-                    effective_owner=row_owner,
-                    effective_ns=row_ns,
-                    in_envelope_index=in_envelope_index,
-                    preserve_owner=preserve_owner,
-                    require_in_envelope=require_in_envelope,
-                    freshly_inserted_uuids=freshly_inserted_version_uuids,
-                )
-                if not ok:
-                    _bump(stats.sidecars_failed, surface)
-                    stats.errors.append(
-                        f"[{surface}] {entry['id']}: parent_version_id/"
-                        f"merge_parents reference foreign-tenant or "
-                        f"foreign-record version(s) {bad}; rejected"
-                    )
-                    failed_record_ids.add(entry["record_id"])
-                    continue
-
-            metadata = entry.get("metadata") or {}
-            # Per-row SAVEPOINT — same reasoning as kg_triples loop.
-            try:
-                async with conn.transaction():
-                    row = await conn.execute(
-                        """
-                        INSERT INTO memory_versions (
-                            id, memory_id, version_num, content,
-                            category, subcategory, metadata, verbatim_content,
-                            owner_id, namespace, permission_mode,
-                            source_model, source_provider, source_session, source_agent,
-                            snapshot_at, snapshot_by, change_type,
-                            commit_hash, parent_version_id, branch, merge_parents
-                        )
-                        VALUES (
-                            $1::uuid, $2, $3, $4,
-                            $5, $6, $7::jsonb, $8,
-                            $9, $10, COALESCE($11, 600),
-                            $12, $13, $14, $15,
-                            COALESCE($16, NOW()), $17, COALESCE($18, 'create'),
-                            $19, $20::uuid, COALESCE($21, 'main'), $22::uuid[]
-                        )
-                        ON CONFLICT (id) DO NOTHING
-                        """,
-                        entry["id"], entry["record_id"], entry["version_num"], entry["content"],
-                        entry.get("category"), entry.get("subcategory"),
-                        json.dumps(metadata), entry.get("verbatim_content"),
-                        row_owner, row_ns, entry.get("permission_mode"),
-                        entry.get("source_model"), entry.get("source_provider"),
-                        entry.get("source_session"), entry.get("source_agent"),
-                        _parse_iso(entry.get("snapshot_at")), entry.get("snapshot_by"),
-                        entry.get("change_type"),
-                        entry.get("commit_hash"), entry.get("parent_version_id"),
-                        entry.get("branch"), entry.get("merge_parents"),
-                    )
-                if row == "INSERT 0 0":
-                    # Conflict-skip: the row already exists in DB.
-                    # Verify EVERY column populated by the INSERT
-                    # matches the envelope's claim. A stale prior-
-                    # lifetime row could otherwise survive a delete
-                    # + re-import on the root migration path
-                    # (Codex round-25 finding). Includes secondary
-                    # fields: category, subcategory, metadata,
-                    # verbatim_content, permission_mode, source_*,
-                    # snapshot_at, snapshot_by, change_type.
-                    existing = await conn.fetchrow(
-                        "SELECT memory_id, owner_id, namespace, "
-                        "version_num, content, commit_hash, "
-                        "parent_version_id::text AS parent_version_id, "
-                        "branch, merge_parents, category, subcategory, "
-                        "metadata, verbatim_content, permission_mode, "
-                        "source_model, source_provider, source_session, "
-                        "source_agent, snapshot_at, snapshot_by, "
-                        "change_type "
-                        "FROM memory_versions WHERE id = $1::uuid",
-                        entry["id"],
-                    )
-                    expected_parent = (
-                        str(entry["parent_version_id"])
-                        if entry.get("parent_version_id") else None
-                    )
-                    expected_branch = entry.get("branch") or "main"
-                    expected_merge_parents = entry.get("merge_parents") or None
-                    expected_change_type = entry.get("change_type") or "create"
-                    expected_permission_mode = entry.get("permission_mode") or 600
-                    actual_merge_parents = (
-                        existing["merge_parents"] if existing else None
-                    )
-
-                    def _norm_mp(mp):
-                        if not mp:
-                            return None
-                        return [str(x) for x in mp]
-
-                    def _norm_jsonb(j):
-                        # asyncpg returns JSONB as str or dict
-                        # depending on codec config; normalize both
-                        # sides through json.loads if str.
-                        if j is None:
-                            return None
-                        if isinstance(j, str):
-                            try:
-                                return json.loads(j)
-                            except Exception:
-                                return j
-                        return j
-
-                    expected_metadata = entry.get("metadata") or {}
-                    actual_metadata = (
-                        _norm_jsonb(existing["metadata"]) if existing else None
-                    )
-                    expected_snapshot_at = _parse_iso(entry.get("snapshot_at"))
-                    actual_snapshot_at = (
-                        existing["snapshot_at"] if existing else None
-                    )
-                    # snapshot_at uses COALESCE($16, NOW()) on
-                    # INSERT (line ~1495). When envelope omits it,
-                    # DB stores NOW(); on retry the envelope still
-                    # has None so strict equality would falsely
-                    # reject the retry (round-38).
-                    #
-                    # Round-39 escalation: tolerance is ONLY safe
-                    # when the referenced memory was NOT freshly
-                    # inserted in this transaction. memory_versions
-                    # intentionally survives memory deletion (audit
-                    # / branch-restore guarantee), so a fresh
-                    # memory hitting INSERT 0 0 on memory_versions
-                    # is a retained prior-lifetime row, NOT an
-                    # idempotent retry. Without the gate, omitted
-                    # snapshot_at would let stale history attach to
-                    # the new memory.
-                    fresh_memory = (
-                        inserted_record_ids is not None
-                        and entry.get("record_id") in inserted_record_ids
-                    )
-                    tolerate_snapshot_at = (
-                        not fresh_memory and entry.get("snapshot_at") is None
-                    )
-                    if existing is None or (
-                        existing["memory_id"] != entry["record_id"]
-                        or existing["owner_id"] != row_owner
-                        or existing["namespace"] != row_ns
-                        or existing["version_num"] != entry["version_num"]
-                        or existing["content"] != entry["content"]
-                        # commit_hash is NOT NULL and INSERT uses
-                        # entry.get("commit_hash") verbatim — so an
-                        # entry that hit conflict MUST have a
-                        # truthy commit_hash. Don't wildcard the
-                        # comparison: an empty/missing commit_hash
-                        # means the entry is malformed, reject it.
-                        or (not entry.get("commit_hash") or
-                            existing["commit_hash"] != entry["commit_hash"])
-                        or existing["parent_version_id"] != expected_parent
-                        or existing["branch"] != expected_branch
-                        or _norm_mp(actual_merge_parents) != _norm_mp(expected_merge_parents)
-                        or existing["category"] != entry.get("category")
-                        or existing["subcategory"] != entry.get("subcategory")
-                        or actual_metadata != expected_metadata
-                        or existing["verbatim_content"] != entry.get("verbatim_content")
-                        or existing["permission_mode"] != expected_permission_mode
-                        or existing["source_model"] != entry.get("source_model")
-                        or existing["source_provider"] != entry.get("source_provider")
-                        or existing["source_session"] != entry.get("source_session")
-                        or existing["source_agent"] != entry.get("source_agent")
-                        # snapshot_at: round-33 enforced IS NOT
-                        # DISTINCT FROM semantics; round-38 added
-                        # COALESCE-tolerance for envelope-NULL since
-                        # the INSERT path stores NOW() in that case.
-                        or (not tolerate_snapshot_at and actual_snapshot_at != expected_snapshot_at)
-                        or existing["snapshot_by"] != entry.get("snapshot_by")
-                        or existing["change_type"] != expected_change_type
-                    ):
-                        _bump(stats.sidecars_failed, surface)
-                        stats.errors.append(
-                            f"[{surface}] {entry['id']}: existing row "
-                            "doesn't match envelope claim (likely "
-                            "prior-lifetime stale row); rejected"
-                        )
-                        failed_record_ids.add(entry["record_id"])
-                        continue
-                    _bump(stats.sidecars_skipped, surface)
-                else:
-                    _bump(stats.sidecars_imported, surface)
-                    # ONLY freshly-inserted (non-conflict) entries
-                    # count as fresh. Conflict-skipped entries
-                    # passed the equality check (so they're
-                    # authorized) but are NOT freshly written by
-                    # this transaction — they could be from a
-                    # prior lifetime that survived. Round-29 fix.
-                    if entry.get("id"):
-                        freshly_inserted_version_uuids.add(str(entry["id"]))
-                # Authorized only after match-verification above
-                # (or successful insert). Stale rows are no longer
-                # treated as safe because of the equality check.
-                authorized_record_ids.add(entry["record_id"])
-                if entry.get("id"):
-                    authorized_version_uuids.add(str(entry["id"]))
-            except Exception as exc:
-                _bump(stats.sidecars_failed, surface)
-                stats.errors.append(f"[{surface}] {entry['id']}: {type(exc).__name__}: {exc}")
-                logger.exception("MPF memory_versions import failed for entry %s", entry.get("id"))
-                failed_record_ids.add(entry["record_id"])
-    return authorized_record_ids, failed_record_ids, authorized_version_uuids
-
-
-async def _import_compression_manifest(
-    conn,
-    sidecar: List[Dict[str, Any]],
-    *,
-    caller_user_id: str,
-    caller_namespace: str,
-    preserve_owner: bool,
-    stats: ImportStats,
-    allowlist: Dict[str, tuple],
-    inserted_record_ids: Optional[set] = None,
-) -> None:
-    """Upsert MPF compression_manifest_entry sidecar entries into
-    memory_compressed_variants. Primary key is `memory_id` (one
-    winner per memory), so re-imports of the same envelope are
-    no-ops via ON CONFLICT (memory_id) DO NOTHING.
-
-    Note: this table has no namespace column — tenancy is owner-
-    only here, intentionally simpler than memory_versions."""
-    surface = "compression_manifest"
-    for entry in sidecar:
-        for required in ("record_id", "engine_id"):
-            if entry.get(required) in (None, ""):
-                _bump(stats.sidecars_failed, surface)
-                stats.errors.append(
-                    f"[{surface}] missing required field {required!r}; skipped"
-                )
-                break
-        else:
-            row_owner, _ = _row_owner_ns(
-                entry,
-                caller_user_id=caller_user_id,
-                caller_namespace=caller_namespace,
-                preserve_owner=preserve_owner,
-                has_namespace_column=False,
-            )
-            # Tenant-scope check. Even though
-            # memory_compressed_variants has no `namespace` column,
-            # the referenced memories row does — and a same-owner
-            # cross-namespace import (alice.ns_A claiming a variant
-            # for alice.ns_B's memory) would silently poison the
-            # compressed content read paths in ns_B. Validate
-            # against the caller's effective namespace, NOT against
-            # row_ns (which is None here because the table has no
-            # namespace column).
-            #
-            # For preserve_owner=true, fall back to the entry's
-            # stated owner_id matching the memory's actual one;
-            # namespace check uses the memory's actual namespace
-            # since the entry has no namespace field by schema.
-            ref_namespace = (
-                None if preserve_owner else caller_namespace
-            )
-            allowed, reason = _is_allowed_reference(
-                entry.get("record_id"),
-                effective_owner=row_owner,
-                effective_namespace=ref_namespace,
-                allowlist=allowlist,
-                # Skip the namespace match only under preserve_owner,
-                # where the caller is root and migrations span ns.
-                require_namespace_match=not preserve_owner,
-            )
-            if not allowed:
-                _bump(stats.sidecars_failed, surface)
-                stats.errors.append(f"[{surface}] {entry.get('record_id')}: {reason}")
-                continue
-            # winner_candidate_id is FK→memory_compression_candidates(id)
-            # ON DELETE SET NULL. The manifest sidecar carries the
-            # winner ID for traceability but CHARON does not also
-            # ship the candidate row, so the FK target may be absent
-            # after import. Pre-validate format in Python (rejecting
-            # malformed UUIDs without a SQL roundtrip), then check
-            # existence INSIDE the per-row SAVEPOINT so a Postgres
-            # error there can't abort the surrounding import
-            # transaction (Codex round-6 finding).
-            #
-            # Lossless on the manifest's primary content (engine_id,
-            # compressed_content, scores) — only the back-pointer
-            # to which contest produced this winner gets nulled when
-            # the candidate row isn't carried.
-            winner_id_raw = entry.get("winner_contest_id")
-            winner_id: Optional[str] = None
-            if winner_id_raw:
-                try:
-                    # uuid.UUID() raises ValueError on malformed input —
-                    # pure Python, no SQL hit.
-                    winner_id = str(uuid.UUID(str(winner_id_raw)))
-                except (ValueError, AttributeError):
-                    winner_id = None
-            # Per-row SAVEPOINT — same reasoning as the other sidecars.
-            try:
-                async with conn.transaction():
-                    if winner_id is not None:
-                        # Tenancy-scoped existence check (Codex round-7
-                        # finding): the candidate row must belong to
-                        # THIS memory_id AND THIS owner. Without that
-                        # constraint, an envelope can point its
-                        # winner_candidate_id at another tenant's
-                        # candidate UUID (if known), creating cross-
-                        # tenant linkage in the audit log. The check
-                        # also runs inside the savepoint so a malformed
-                        # input or DB error rolls back only this row.
-                        exists = await conn.fetchval(
-                            "SELECT 1 FROM memory_compression_candidates "
-                            "WHERE id = $1::uuid AND memory_id = $2 "
-                            "AND owner_id = $3",
-                            winner_id, entry["record_id"], row_owner,
-                        )
-                        if not exists:
-                            winner_id = None
-                    row = await conn.execute(
-                        """
-                        INSERT INTO memory_compressed_variants (
-                            memory_id, owner_id, winner_candidate_id,
-                            engine_id, engine_version, compressed_content,
-                            compressed_tokens, compression_ratio,
-                            quality_score, composite_score,
-                            scoring_profile, judge_model, selected_at
-                        )
-                        VALUES (
-                            $1, $2, $3::uuid,
-                            $4, $5, $6,
-                            $7, $8,
-                            $9, $10,
-                            COALESCE($11, 'balanced'), $12,
-                            COALESCE($13, NOW())
-                        )
-                        ON CONFLICT (memory_id) DO NOTHING
-                        """,
-                        entry["record_id"], row_owner, winner_id,
-                        entry["engine_id"], entry.get("engine_version"),
-                        entry.get("compressed_content"),
-                        entry.get("compressed_tokens"),
-                        entry.get("compression_ratio"),
-                        entry.get("quality_score"),
-                        entry.get("composite_score"),
-                        entry.get("scoring_profile"),
-                        entry.get("judge_model"),
-                        _parse_iso(entry.get("selected_at")),
-                    )
-                if row == "INSERT 0 0":
-                    # Exact-match conflict verification (round 32 —
-                    # mirrors round-23 / round-30 fixes for
-                    # memory_versions and kg_triples). The PK is
-                    # memory_id, so a re-import of a corrected
-                    # variant with the same memory_id would silently
-                    # skip and leave stale compressed_content in
-                    # place. Read paths use this content directly,
-                    # so the staleness is operator-visible.
-                    existing = await conn.fetchrow(
-                        "SELECT owner_id, winner_candidate_id::text "
-                        "AS winner_candidate_id, engine_id, "
-                        "engine_version, compressed_content, "
-                        "compressed_tokens, compression_ratio, "
-                        "quality_score, composite_score, "
-                        "scoring_profile, judge_model, selected_at "
-                        "FROM memory_compressed_variants "
-                        "WHERE memory_id = $1",
-                        entry["record_id"],
-                    )
-                    expected_winner = (
-                        str(winner_id) if winner_id else None
-                    )
-                    expected_scoring = entry.get("scoring_profile") or "balanced"
-                    expected_selected_at = _parse_iso(entry.get("selected_at"))
-                    # selected_at uses COALESCE($13, NOW()) on
-                    # INSERT (line ~1760). When envelope omits it,
-                    # DB stores NOW(); on retry the envelope still
-                    # has None so strict equality would falsely
-                    # reject the retry (round-38).
-                    #
-                    # Round-39 escalation: tolerance is ONLY safe
-                    # when the referenced memory was NOT freshly
-                    # inserted in this transaction. Without the
-                    # gate, omitted selected_at would let a stale
-                    # prior-lifetime variant attach to a new memory
-                    # with the same record_id.
-                    fresh_memory = (
-                        inserted_record_ids is not None
-                        and entry.get("record_id") in inserted_record_ids
-                    )
-                    tolerate_selected_at = (
-                        not fresh_memory and entry.get("selected_at") is None
-                    )
-                    if existing is None or (
-                        existing["owner_id"] != row_owner
-                        or existing["winner_candidate_id"] != expected_winner
-                        or existing["engine_id"] != entry["engine_id"]
-                        or existing["engine_version"] != entry.get("engine_version")
-                        or existing["compressed_content"] != entry.get("compressed_content")
-                        or existing["compressed_tokens"] != entry.get("compressed_tokens")
-                        or existing["compression_ratio"] != entry.get("compression_ratio")
-                        or existing["quality_score"] != entry.get("quality_score")
-                        or existing["composite_score"] != entry.get("composite_score")
-                        or existing["scoring_profile"] != expected_scoring
-                        or existing["judge_model"] != entry.get("judge_model")
-                        # selected_at: round-33 enforced IS NOT
-                        # DISTINCT FROM semantics; round-38 added
-                        # COALESCE-tolerance for envelope-NULL since
-                        # the INSERT stores NOW() in that case.
-                        or (not tolerate_selected_at and existing["selected_at"] != expected_selected_at)
-                    ):
-                        _bump(stats.sidecars_failed, surface)
-                        stats.errors.append(
-                            f"[{surface}] {entry.get('record_id')}: "
-                            "existing variant doesn't match envelope "
-                            "claim (likely stale prior-lifetime row); "
-                            "rejected"
-                        )
-                        continue
-                    _bump(stats.sidecars_skipped, surface)
-                else:
-                    _bump(stats.sidecars_imported, surface)
-            except Exception as exc:
-                _bump(stats.sidecars_failed, surface)
-                stats.errors.append(
-                    f"[{surface}] {entry.get('record_id')}: {type(exc).__name__}: {exc}"
-                )
-                logger.exception(
-                    "MPF compression_manifest import failed for record_id %s",
-                    entry.get("record_id"),
-                )
 
 
 @router.post("/import", response_model=ImportStats, status_code=200)
 async def import_memories(
-    envelope: MPFEnvelope = Body(..., description="An MPF v0.1 envelope."),
-    preserve_owner: bool = Query(
-        False,
-        description=(
-            "Root only. When true, honor the owner_id + namespace on "
-            "each incoming record instead of rewriting to the caller's "
-            "identity. Required for cross-tenant migrations; refused for "
-            "non-root callers even if passed."
-        ),
-    ),
+    envelope: Annotated[MPFEnvelope, Body()],
+    preserve_owner: Annotated[bool, Query()] = False,
     user: UserContext = Depends(get_current_user),
 ):
-    """Import an MPF envelope.
-
-    Accepts `kind: memory` records and the three MNEMOS-native
-    sidecars (`kg_triples`, `memory_versions`,
-    `compression_manifest`). Other record kinds (document, fact,
-    event) are counted under `unsupported_kinds` and skipped per
-    the spec's forward-compatibility rule — they need per-adapter
-    payload mapping that's deferred to a follow-up.
-
-    All sidecar inserts run in the same transaction as the
-    records, so a fatal error on any surface rolls the whole
-    envelope back."""
     if not _lc._pool:
         raise HTTPException(status_code=503, detail="Database pool not available")
 
-    if not envelope.mpf_version.startswith(MPF_VERSION_PREFIX):
-        # Forward-compat ratchet: accept any 0.1.x envelope. Newer
-        # minor versions add optional fields that this server may
-        # ignore, but the required-fields contract is unchanged
-        # within the 0.1 series. A 0.2.x bump would change required
-        # contracts and re-introduces strict matching.
-        raise HTTPException(
-            status_code=415,
-            detail=f"Unsupported MPF version {envelope.mpf_version!r}; expected {MPF_VERSION_PREFIX}x",
-        )
-
-    if preserve_owner and not is_root(user):
-        raise HTTPException(
-            status_code=403, detail="preserve_owner=true requires root",
-        )
-
-    stats = ImportStats(
-        imported=0, skipped=0, failed=0,
-        unsupported_kinds={},
-        sidecars_imported={}, sidecars_skipped={}, sidecars_failed={},
-        errors=[],
-    )
-
-    # Restrict trigger-suppressed memory_versions imports to the
-    # root + preserve_owner=true admin/migration path. CHARON v0.2
-    # adversarial review (Codex rounds 16-24) systematically
-    # surfaced a sequence of stale-state edge cases on the non-root
-    # path arising from the interaction of (a) caller-scoped
-    # deterministic id derivation, (b) ON CONFLICT idempotency,
-    # (c) memory_versions surviving memory deletion (audit feature),
-    # and (d) trigger suppression for sidecar imports. Each finding
-    # was a real bypass with a local fix, but the underlying state
-    # machine kept producing new variants — every persisted column
-    # had to be in an equality check, every prior-lifetime artifact
-    # had to be cleared, and the contract committed us to keeping
-    # those checks exhaustive as the schema evolved.
-    #
-    # Architectural restriction: the documented migration path is
-    # root + preserve_owner=true (`--preserve-metadata` on the CLI),
-    # which already requires a root bearer token and skips id
-    # derivation. The non-root path was never load-bearing for the
-    # documented migration or interop use cases — non-root callers
-    # who want authoritative version history can use the records-
-    # only import and rely on the trigger-fired default v1, or
-    # ship kg_triples / compression_manifest sidecars (no trigger
-    # interaction).
-    if envelope.memory_versions and not (preserve_owner and is_root(user)):
-        raise HTTPException(
-            status_code=403,
-            detail=(
-                "memory_versions sidecar import requires root + "
-                "preserve_owner=true (the admin/migration path; "
-                "use --preserve-metadata in tools/memory_import.py "
-                "with a root bearer token). Non-root callers can "
-                "import records and rely on the trigger-fired "
-                "default v1 history, or ship kg_triples / "
-                "compression_manifest sidecars without restriction. "
-                "The non-root + memory_versions sidecar combination "
-                "is not supported under CHARON v0.2 due to "
-                "deterministic-id stale-state interactions."
-            ),
-        )
-
-    # Pre-validate memory_versions coverage BEFORE opening the
-    # transaction. When an envelope carries memory_versions, the
-    # records loop runs with the version-snapshot trigger
-    # suppressed — so any record_id that *isn't* covered by the
-    # sidecar would land in `memories` without a v1 history entry
-    # at all. That's worse than the original collision bug:
-    # silently-unversioned production data. Refuse the envelope
-    # up front rather than partially-import broken history.
-    if envelope.memory_versions:
-        memory_record_ids = {
-            r.id for r in envelope.records if r.kind == "memory"
-        }
-        sidecar_versioned_ids = {
-            e.get("record_id") for e in envelope.memory_versions
-            if e.get("record_id")
-        }
-        uncovered = memory_record_ids - sidecar_versioned_ids
-        if uncovered:
-            raise HTTPException(
-                status_code=400,
-                detail=(
-                    "memory_versions sidecar must cover every "
-                    f"kind: memory record being imported. {len(uncovered)} "
-                    f"record(s) have no version entry: "
-                    f"{sorted(uncovered)[:5]}{'...' if len(uncovered) > 5 else ''}. "
-                    "Either ship a complete sidecar or omit the "
-                    "memory_versions array entirely (the trigger will "
-                    "synthesize default v1 history)."
-                ),
-            )
-
     async with _lc._pool.acquire() as conn:
-        async with conn.transaction():
-            # When the envelope carries its own memory_versions
-            # sidecar, the import is restoring authoritative history.
-            # The mnemos_version_snapshot trigger would otherwise
-            # synthesize a fresh v1 on every memory INSERT, then
-            # collide with the envelope's v1 on the
-            # idx_mv_main_linear partial unique index `(memory_id,
-            # version_num) WHERE branch='main'`. Suppress the trigger
-            # for the duration of this transaction; the envelope IS
-            # the version log. Pre-validation above guarantees that
-            # every record being inserted has authoritative history.
-            #
-            # Targeted suppression via the `mnemos.suppress_version_snapshot`
-            # custom GUC. The trigger creation in
-            # db/migrations_charon_trigger_guard.sql attaches a
-            # WHEN clause that no-ops the three version-snapshot
-            # triggers when this GUC is '1'. Custom dot-namespaced
-            # GUCs are settable per-transaction by any role — no
-            # superuser required, unlike `session_replication_role`,
-            # which is what the v3.3 cut originally used. The GUC also
-            # leaves FK enforcement and every other user-defined
-            # trigger untouched.
-            if envelope.memory_versions:
-                await conn.execute(
-                    "SET LOCAL mnemos.suppress_version_snapshot = '1'"
-                )
-
-            # Track which memory IDs this import actually INSERTed
-            # (vs which hit ON CONFLICT DO NOTHING and were
-            # pre-existing). The post-import v1 verification scopes
-            # to THIS set, not envelope.records — otherwise a
-            # pre-existing memory with no v1 history would
-            # roll back an unrelated import (Codex round-4 finding).
-            inserted_record_ids: set = set()
-
-            # Persisted IDs the envelope CLAIMED to define but whose
-            # insert was rejected (conflict-row content/owner/ns
-            # mismatch, or per-row exception). Used to actively
-            # exclude these from the sidecar allowlist after the
-            # records loop — without this, a same-tenant collision
-            # where the existing DB row holds different content
-            # would still let kg_triples / compression_manifest
-            # attach to the stale row, since the row legitimately
-            # exists in DB and would otherwise pass the allowlist
-            # SELECT (Codex round-35 finding).
-            rejected_persisted_ids: set = set()
-
-            # Per-request id remap: envelope record_id -> the id the
-            # row was actually persisted under. For non-root callers
-            # (no preserve_owner) we DERIVE a caller-scoped id from
-            # caller identity + envelope id + content, so an attacker
-            # can't probe for foreign memory_ids via
-            # imported-vs-skipped outcome inference. Sidecar
-            # references in the same envelope get translated through
-            # this dict before allowlist + INSERT (Codex round-14
-            # finding). For root + preserve_owner=true, the dict is
-            # the identity mapping — admin migration preserves ids.
-            id_remap: Dict[str, str] = {}
-            non_root_id_rewrite = not (preserve_owner and is_root(user))
-
-            for record in envelope.records:
-                if record.kind != "memory":
-                    stats.unsupported_kinds[record.kind] = (
-                        stats.unsupported_kinds.get(record.kind, 0) + 1
-                    )
-                    continue
-
-                if record.payload_version != MEMORY_PAYLOAD_VERSION:
-                    # Payload version mismatch isn't fatal — record the
-                    # skip for operator visibility. Migrating payloads
-                    # across versions is a follow-up commit.
-                    stats.skipped += 1
-                    stats.errors.append(
-                        f"{record.id}: unsupported payload_version "
-                        f"{record.payload_version!r}; expected {MEMORY_PAYLOAD_VERSION}"
-                    )
-                    # Codex round-37: early-continue paths happen
-                    # BEFORE persisted_id is computed and id_remap
-                    # is populated. Without recording the envelope
-                    # id as rejected, root+preserve_owner sidecars
-                    # referencing the same id would still attach
-                    # to whatever pre-existing DB row matches,
-                    # because the allowlist build sees nothing
-                    # rejected. Add record.id to
-                    # rejected_persisted_ids — under root+preserve
-                    # persisted_id would have been record.id so
-                    # this is the right key to strip from the
-                    # allowlist post-build.
-                    rejected_persisted_ids.add(record.id)
-                    continue
-
-                p = record.payload
-
-                if preserve_owner:
-                    imported_owner = p.get("owner_id") or user.user_id
-                    imported_ns = p.get("namespace") or user.namespace
-                else:
-                    imported_owner = user.user_id
-                    imported_ns = user.namespace
-
-                content = p.get("content")
-                if not content or not str(content).strip():
-                    stats.failed += 1
-                    stats.errors.append(f"{record.id}: empty content; skipped")
-                    # Same hazard as the payload_version-mismatch
-                    # branch above (Codex round-37): an envelope
-                    # record that fails empty-content validation
-                    # must not authorize sidecars referencing the
-                    # same id against any pre-existing DB row.
-                    rejected_persisted_ids.add(record.id)
-                    continue
-
-                category = p.get("category") or "imported"
-                subcategory = p.get("subcategory")
-                permission_mode = p.get("permission_mode") or 600
-                metadata = p.get("metadata") or {}
-                quality_rating = p.get("quality_rating") or 75
-
-                # Determine the persisted id. Root with
-                # preserve_owner=true keeps envelope id verbatim
-                # (admin migration); non-root derives a caller-
-                # scoped hash so foreign-id collisions don't
-                # leak existence (round-14 fix).
-                if non_root_id_rewrite:
-                    persisted_id = _derive_caller_scoped_id(
-                        record.id,
-                        caller_owner=imported_owner,
-                        caller_namespace=imported_ns,
-                        content=str(content),
-                    )
-                else:
-                    persisted_id = record.id
-                id_remap[record.id] = persisted_id
-
-                # ON CONFLICT DO NOTHING gives us idempotent re-imports —
-                # running /v1/export followed by /v1/import against the
-                # same DB is a no-op. For non-root, idempotency is on
-                # the derived id (which collapses identical content
-                # under the same caller into one row).
-                #
-                # Wrap the per-row INSERT in a SAVEPOINT (asyncpg's
-                # nested transaction context) so a Postgres-level
-                # error on one row aborts ONLY that row, not the
-                # whole import transaction. Without this, a single
-                # constraint violation poisons the outer transaction
-                # and every subsequent statement (allowlist SELECT,
-                # sidecar INSERTs, post-verification SELECT) fails
-                # with InFailedSQLTransaction. Codex round-5 finding.
-                try:
-                    async with conn.transaction():
-                        row = await conn.execute(
-                            """
-                            INSERT INTO memories (
-                                id, content, category, subcategory, metadata,
-                                quality_rating, owner_id, namespace, permission_mode,
-                                source_model, source_provider, source_session, source_agent,
-                                created, updated
-                            )
-                            VALUES (
-                                $1, $2, $3, $4, $5::jsonb,
-                                $6, $7, $8, $9,
-                                $10, $11, $12, $13,
-                                COALESCE($14, NOW()), COALESCE($15, NOW())
-                            )
-                            ON CONFLICT (id) DO NOTHING
-                            """,
-                            persisted_id, content, category, subcategory,
-                            json.dumps(metadata),
-                            quality_rating, imported_owner, imported_ns, permission_mode,
-                            p.get("source_model"), p.get("source_provider"),
-                            p.get("source_session"), p.get("source_agent"),
-                            # memories.created/updated are TIMESTAMP
-                            # (no tz) — pre-strip to naive UTC so
-                            # asyncpg doesn't apply session-tz shift
-                            # (Codex round-17 finding).
-                            _parse_iso_naive(p.get("created")),
-                            _parse_iso_naive(p.get("updated")),
-                        )
-                    if row == "INSERT 0 0":
-                        # Conflict-skip on memories table: verify
-                        # the existing row matches the envelope's
-                        # claim before treating it as authorized
-                        # for sidecar attachment. Otherwise a
-                        # root+preserve import into a non-empty
-                        # target with a same-id-different-content
-                        # memory could land sidecars (kg_triples /
-                        # compression_manifest) against the stale
-                        # memory body. Round-34 finding.
-                        #
-                        # Round-36 finding extended this: the prior
-                        # check only compared content/owner/namespace,
-                        # leaving stale permission_mode (federation
-                        # visibility), category, metadata,
-                        # quality_rating, source_*, subcategory able
-                        # to authorize sidecar attachment. Now we
-                        # compare every envelope-bound column. For
-                        # the COALESCE'd timestamp fields
-                        # (created/updated) we tolerate a non-match
-                        # only when the envelope didn't provide a
-                        # value — because the prior INSERT used
-                        # NOW() in that case and re-imports without
-                        # those fields are legitimately idempotent.
-                        existing_mem = await conn.fetchrow(
-                            "SELECT content, category, subcategory, "
-                            "metadata, quality_rating, owner_id, "
-                            "namespace, permission_mode, "
-                            "source_model, source_provider, "
-                            "source_session, source_agent, "
-                            "created, updated "
-                            "FROM memories WHERE id = $1",
-                            persisted_id,
-                        )
-                        envelope_metadata_json = json.dumps(
-                            metadata, sort_keys=True,
-                        )
-                        existing_metadata_json = (
-                            json.dumps(
-                                existing_mem["metadata"]
-                                if isinstance(
-                                    existing_mem["metadata"], dict,
-                                )
-                                else json.loads(
-                                    existing_mem["metadata"] or "{}"
-                                ),
-                                sort_keys=True,
-                            )
-                            if existing_mem is not None
-                            else None
-                        )
-                        envelope_created = _parse_iso_naive(
-                            p.get("created")
-                        )
-                        envelope_updated = _parse_iso_naive(
-                            p.get("updated")
-                        )
-                        mismatched_fields: list[str] = []
-                        if existing_mem is None:
-                            mismatched_fields.append("row missing")
-                        else:
-                            checks = [
-                                ("content", existing_mem["content"], content),
-                                ("category", existing_mem["category"], category),
-                                ("subcategory", existing_mem["subcategory"], subcategory),
-                                ("metadata", existing_metadata_json, envelope_metadata_json),
-                                ("quality_rating", existing_mem["quality_rating"], quality_rating),
-                                ("owner_id", existing_mem["owner_id"], imported_owner),
-                                ("namespace", existing_mem["namespace"], imported_ns),
-                                ("permission_mode", existing_mem["permission_mode"], permission_mode),
-                                ("source_model", existing_mem["source_model"], p.get("source_model")),
-                                ("source_provider", existing_mem["source_provider"], p.get("source_provider")),
-                                ("source_session", existing_mem["source_session"], p.get("source_session")),
-                                ("source_agent", existing_mem["source_agent"], p.get("source_agent")),
-                            ]
-                            for col, db_val, env_val in checks:
-                                if db_val != env_val:
-                                    mismatched_fields.append(col)
-                            # COALESCE-tolerance: only assert
-                            # timestamp match when envelope provided
-                            # a value. Without this, every legitimate
-                            # idempotent re-import that omits
-                            # created/updated would fail because the
-                            # original INSERT used NOW().
-                            if envelope_created is not None and existing_mem["created"] != envelope_created:
-                                mismatched_fields.append("created")
-                            if envelope_updated is not None and existing_mem["updated"] != envelope_updated:
-                                mismatched_fields.append("updated")
-
-                        if mismatched_fields:
-                            stats.failed += 1
-                            stats.errors.append(
-                                f"{record.id}: existing memory row "
-                                "doesn't match envelope payload "
-                                f"({', '.join(mismatched_fields)} differ); "
-                                "sidecar attachment refused"
-                            )
-                            # Codex round-35 finding: id_remap was
-                            # populated BEFORE the conflict check
-                            # (line 2062), so leaving it in place
-                            # would let sidecar references rewrite
-                            # to persisted_id and attach to the
-                            # stale DB row via the allowlist.
-                            # Remove the remap entry AND mark the
-                            # persisted_id rejected so the allowlist
-                            # build excludes it.
-                            id_remap.pop(record.id, None)
-                            rejected_persisted_ids.add(persisted_id)
-                            continue
-                        stats.skipped += 1
-                    else:
-                        stats.imported += 1
-                        inserted_record_ids.add(persisted_id)
-                except Exception as exc:
-                    stats.failed += 1
-                    stats.errors.append(f"{record.id}: {type(exc).__name__}: {exc}")
-                    logger.exception("MPF import failed for record %s", record.id)
-                    # Same hazard as the conflict-mismatch branch:
-                    # id_remap was set before the INSERT ran. If a
-                    # per-row exception aborted the SAVEPOINT, the
-                    # row didn't land — but a same-id row may still
-                    # exist in DB from a prior import. Drop the
-                    # remap and reject the persisted_id so sidecars
-                    # can't ride a partial-failure into the stale
-                    # row (Codex round-35 finding extended).
-                    id_remap.pop(record.id, None)
-                    rejected_persisted_ids.add(persisted_id)
-
-            # Build the per-request allowlist of memory_ids the
-            # caller may attach sidecars to. Computed AFTER the
-            # records loop so any memory just imported in this
-            # request is included. The check inside each helper
-            # confirms post-rewrite owner+namespace matches the
-            # memory's actual ownership in DB — so an envelope
-            # cannot smuggle sidecars onto another tenant's
-            # memory_id by labeling them with the caller's
-            # identity. This is the cross-tenant-attachment fix.
-            # Translate sidecar memory references through the
-            # records-loop id remap. Under non-root, envelope ids
-            # were rewritten to caller-scoped hashes; sidecar
-            # references must follow so they bind to the right
-            # memories. Untouched ids (sidecar references that
-            # weren't in envelope.records) are passed through —
-            # they'll be rejected by the allowlist gate downstream
-            # if they don't belong to the caller.
-            if id_remap:
-                for entry in envelope.kg_triples or []:
-                    mid = entry.get("memory_id")
-                    if mid and mid in id_remap:
-                        entry["memory_id"] = id_remap[mid]
-                for entry in envelope.memory_versions or []:
-                    rid = entry.get("record_id")
-                    if rid and rid in id_remap:
-                        entry["record_id"] = id_remap[rid]
-                for entry in envelope.compression_manifest or []:
-                    rid = entry.get("record_id")
-                    if rid and rid in id_remap:
-                        entry["record_id"] = id_remap[rid]
-
-            # Scope the allowlist SELECT structurally so a non-root
-            # caller can't probe for foreign memory_ids via
-            # rejection-message inference (Codex round-13 finding).
-            # Non-root: SELECT is filtered to caller's owner+ns;
-            # foreign ids simply don't appear, identical to
-            # nonexistent ids. Root with preserve_owner=true: lookup
-            # is unscoped — that's the migration/admin path where
-            # cross-tenant resolution is the intended behavior.
-            if is_root(user) and preserve_owner:
-                scope_owner: Optional[str] = None
-                scope_namespace: Optional[str] = None
-            else:
-                scope_owner = user.user_id
-                scope_namespace = user.namespace
-            allowlist = await _build_referenced_memory_allowlist(
-                conn, envelope,
-                scope_owner=scope_owner,
-                scope_namespace=scope_namespace,
-            )
-            # Codex round-35 finding: even with id_remap cleared
-            # for rejected records, a sidecar that references the
-            # persisted_id directly (envelope record_id == DB id
-            # under root + preserve_owner) would still resolve in
-            # the unscoped allowlist SELECT and attach to the
-            # stale row. Strip the rejected ids from the allowlist
-            # so any sidecar referencing them is refused at
-            # _is_allowed_reference time.
-            for rid in rejected_persisted_ids:
-                allowlist.pop(rid, None)
-
-            # Sidecars run inside the same transaction so a partial
-            # failure rolls everything back. Order: kg_triples first
-            # (no FK to memories required), then memory_versions and
-            # compression_manifest, both of which reference memories.id.
-            if envelope.kg_triples:
-                await _import_kg_triples(
-                    conn, envelope.kg_triples,
-                    caller_user_id=user.user_id,
-                    caller_namespace=user.namespace,
-                    preserve_owner=preserve_owner,
-                    stats=stats,
-                    allowlist=allowlist,
-                    inserted_record_ids=inserted_record_ids,
-                )
-            if envelope.memory_versions:
-                (
-                    authorized_version_ids,
-                    failed_version_record_ids,
-                    authorized_version_uuids,
-                ) = await _import_memory_versions(
-                    conn, envelope.memory_versions,
-                    caller_user_id=user.user_id,
-                    caller_namespace=user.namespace,
-                    preserve_owner=preserve_owner,
-                    stats=stats,
-                    allowlist=allowlist,
-                    inserted_record_ids=inserted_record_ids,
-                )
-                # All-or-nothing per record (Codex round-7 finding):
-                # if a record was INSERTed in this transaction AND
-                # any of its memory_versions sidecar entries failed,
-                # roll back the whole import. Otherwise the record
-                # commits with partial / inconsistent history and
-                # the API returns a 200 with sidecars_failed >0,
-                # which is a silent integrity violation.
-                fatal_record_ids = inserted_record_ids & failed_version_record_ids
-                if fatal_record_ids:
-                    sample = sorted(fatal_record_ids)[:5]
-                    extra = "..." if len(fatal_record_ids) > 5 else ""
-                    raise HTTPException(
-                        status_code=500,
-                        detail=(
-                            "CHARON import: memory_versions sidecar had "
-                            f"failed entries for {len(fatal_record_ids)} "
-                            f"newly inserted record(s): {sample}{extra}. "
-                            "Authoritative history is all-or-nothing per "
-                            "record under trigger suppression — partial "
-                            "history would be inconsistent. Transaction "
-                            "rolled back; fix the sidecar and retry."
-                        ),
-                    )
-
-                # Restore memory_branches HEAD pointers. The trigger
-                # would normally do this on memory INSERT, but it's
-                # suppressed during CHARON imports. Without this,
-                # the imported version rows are orphans from the
-                # branch-walk perspective (DAG endpoints return empty).
-                #
-                # Only restore branches for record_ids the import
-                # was authorized to write — rejected cross-tenant
-                # entries must not drive memory_branches mutations
-                # for memories the caller can't touch.
-                # Clear stale memory_branches rows for newly
-                # inserted records before restoring. memory_branches
-                # has no FK cascade with memories, so prior-lifetime
-                # branch rows can survive a delete + re-import even
-                # if memory_versions is now scoped to authorized
-                # UUIDs — the stale branch row would still point at
-                # a stale (or now-stranded) head and let the post-
-                # verification accept it (Codex round-22 finding).
-                # Wipe the slate per inserted record_id, then
-                # restore only authorized branches.
-                if inserted_record_ids:
-                    await conn.execute(
-                        "DELETE FROM memory_branches "
-                        "WHERE memory_id = ANY($1::text[])",
-                        list(inserted_record_ids),
-                    )
-                if authorized_version_ids:
-                    await _restore_memory_branches(
-                        conn, list(authorized_version_ids),
-                        # Scope HEAD selection to THIS request's
-                        # version UUIDs only — stale rows from a
-                        # prior lifetime of the same memory_id
-                        # (post-delete + re-import via deterministic
-                        # id) cannot become the new HEAD.
-                        authorized_version_uuids=list(authorized_version_uuids),
-                    )
-            if envelope.compression_manifest:
-                await _import_compression_manifest(
-                    conn, envelope.compression_manifest,
-                    caller_user_id=user.user_id,
-                    caller_namespace=user.namespace,
-                    preserve_owner=preserve_owner,
-                    stats=stats,
-                    allowlist=allowlist,
-                    inserted_record_ids=inserted_record_ids,
-                )
-
-            # Post-import v1 verification: every memory THIS request
-            # actually INSERTed (under trigger suppression) must
-            # have at least one row in memory_versions. Pre-existing
-            # rows that hit ON CONFLICT DO NOTHING are NOT verified
-            # here — their v1 history is whatever it already was;
-            # this transaction didn't touch them.
-            if envelope.memory_versions:
-                # Coverage check: every newly INSERTed record must
-                # have at least one version row. (Pre-existing
-                # memories aren't checked here — their version
-                # history was set up by their original import.)
-                if inserted_record_ids:
-                    covered = await conn.fetch(
-                        "SELECT DISTINCT memory_id FROM memory_versions "
-                        "WHERE memory_id = ANY($1::text[])",
-                        list(inserted_record_ids),
-                    )
-                    covered_ids = {r["memory_id"] for r in covered}
-                    uncovered = inserted_record_ids - covered_ids
-                    if uncovered:
-                        sample = sorted(uncovered)[:5]
-                        extra = "..." if len(uncovered) > 5 else ""
-                        raise HTTPException(
-                            status_code=500,
-                            detail=(
-                                "CHARON import inserted "
-                                f"{len(uncovered)} memory record(s) "
-                                "without version history under "
-                                f"trigger suppression: {sample}{extra}. "
-                                "Sidecar likely contained malformed or "
-                                "rejected entries that did not produce "
-                                "rows. Transaction rolled back."
-                            ),
-                        )
-
-                # HEAD-content integrity check: every record_id
-                # whose branch HEAD was touched (newly inserted OR
-                # pre-existing memory whose sidecar landed) must
-                # have memories.content == HEAD memory_versions
-                # content. Otherwise /log + branch traversal report
-                # forged history (Codex round-18 + round-19 findings).
-                # Round-19: extend to authorized_version_ids, not
-                # just inserted_record_ids — sidecar-only imports
-                # against pre-existing memories can poison HEAD too.
-                # authorized_version_ids was populated earlier when
-                # envelope.memory_versions was processed; it
-                # represents every record_id whose sidecar
-                # successfully imported (or hit ON CONFLICT against
-                # an authorized pre-existing row).
-                touched_ids = inserted_record_ids | authorized_version_ids
-                if touched_ids:
-                    # LEFT JOIN so records WITHOUT a main branch
-                    # row come back with NULL head_content. Round-20
-                    # finding: an envelope can ship memory_versions
-                    # only on branch='feature', land it, and skip
-                    # this check entirely under the previous INNER
-                    # JOIN — the memory commits without any main-
-                    # branch history.
-                    head_check = await conn.fetch(
-                        """
-                        SELECT m.id, m.content AS memory_content,
-                               mv.content AS head_content
-                        FROM memories m
-                        LEFT JOIN memory_branches b
-                          ON b.memory_id = m.id AND b.name = 'main'
-                        LEFT JOIN memory_versions mv
-                          ON mv.id = b.head_version_id
-                        WHERE m.id = ANY($1::text[])
-                        """,
-                        list(touched_ids),
-                    )
-                    # NEW inserted records MUST have a main HEAD
-                    # whose content matches. Pre-existing records
-                    # that the import only sidecar-touched may not
-                    # have had main updated by this request, so
-                    # missing main is allowed for them iff the
-                    # missing case means we never reached them
-                    # (their existing main is whatever it was
-                    # before, and we didn't touch it).
-                    missing_inserted = []
-                    divergent = []
-                    in_db_inserted = inserted_record_ids
-                    for r in head_check:
-                        rid = r["id"]
-                        head_content = r["head_content"]
-                        memory_content = r["memory_content"]
-                        if head_content is None:
-                            if rid in in_db_inserted:
-                                missing_inserted.append(rid)
-                            # else: pre-existing record without
-                            # main branch; the records loop didn't
-                            # touch it so existing state stands.
-                        elif memory_content != head_content:
-                            divergent.append(rid)
-                    if missing_inserted:
-                        sample = sorted(missing_inserted)[:5]
-                        extra = "..." if len(missing_inserted) > 5 else ""
-                        raise HTTPException(
-                            status_code=500,
-                            detail=(
-                                "CHARON import inserted "
-                                f"{len(missing_inserted)} memory "
-                                "record(s) with no main-branch HEAD: "
-                                f"{sample}{extra}. The envelope's "
-                                "memory_versions sidecar must include "
-                                "a branch='main' entry for every "
-                                "kind:memory record being imported. "
-                                "Transaction rolled back."
-                            ),
-                        )
-                    if divergent:
-                        sample = sorted(divergent)[:5]
-                        extra = "..." if len(divergent) > 5 else ""
-                        raise HTTPException(
-                            status_code=500,
-                            detail=(
-                                "CHARON import: live memory content "
-                                "diverges from restored memory_versions "
-                                f"HEAD for {len(divergent)} record(s): "
-                                f"{sample}{extra}. The envelope's "
-                                "memory_versions sidecar must include "
-                                "an entry whose content matches each "
-                                "touched memory's content. "
-                                "Transaction rolled back."
-                            ),
-                        )
-
-    logger.info(
-        "[MPF] import: user=%s imported=%d skipped=%d failed=%d unsupported=%s sidecars_imported=%s",
-        user.user_id, stats.imported, stats.skipped, stats.failed,
-        stats.unsupported_kinds, stats.sidecars_imported,
-    )
-    return stats
+        return await _import_memories(
+            conn,
+            envelope=envelope,
+            preserve_owner=preserve_owner,
+            user=user,
+        )
